refactor: split channel_list.py into workers + mixin modules
Split the 5076-line monolithic channel_list.py into focused modules:

- workers.py: Worker classes (M3U/Xtream/STB loaders, link creator),
  tree widget items, SetProviderThread, and Roman numeral helpers
- mixins/playback_mixin.py: Playback, auto-play, resume, VLC/MPV
  launch, and watch history
- mixins/content_loading_mixin.py: Content loading from M3U/Xtream/STB
  providers, worker callbacks, and progress tracking
- mixins/export_mixin.py: Export and save channel list functionality
- mixins/display_mixin.py: Content display, filtering, info panel,
  favorites, and logo loading
- channel_list.py: Core ChannelList class (1208 lines) with init,
  UI setup, navigation, provider management, menus, and shortcuts

The ChannelList class uses multiple inheritance from the mixin classes,
preserving all existing behavior while improving maintainability.
