+++ channel_list.py
@@ -1,34 +1,24 @@
-import base64
-from datetime import datetime
-import html
-import logging
-import os
-from pathlib import Path
-import platform
-import re
-import shutil
-import subprocess
-import time
-from typing import Any, Dict, List, Optional
-from urllib.parse import quote as url_quote
+"""Core ChannelList window - UI setup, navigation, providers, menus, shortcuts.
 
-from PySide6.QtCore import QBuffer, QEvent, QObject, QRect, QSize, Qt, QThread, QTimer, Signal
-from PySide6.QtGui import QAction, QColor, QFont, QFontMetrics, QIcon, QKeySequence, QPixmap
+Domain-specific logic is in mixin modules (mixins/) and worker classes (workers.py).
+"""
+
+from datetime import datetime
+import logging
+from typing import Any, Dict, List, Optional
+
+from PySide6.QtCore import QEvent, Qt, QTimer
+from PySide6.QtGui import QAction, QColor, QFont, QKeySequence
 from PySide6.QtWidgets import (
     QApplication,
     QComboBox,
     QDialog,
-    QFileDialog,
     QHBoxLayout,
     QLabel,
-    QLineEdit,
     QListWidget,
-    QListWidgetItem,
     QMainWindow,
-    QMenu,
     QMessageBox,
     QProgressBar,
-    QProgressDialog,
     QPushButton,
     QSizePolicy,
     QSplitter,
@@ -37,803 +27,28 @@ from PySide6.QtWidgets import (
     QVBoxLayout,
     QWidget,
 )
-import requests
 import tzlocal
-from urlobject import URLObject
-
-from services.provider_api import (
-    base_from_url,
-    stb_request_url,
-    xtream_choose_resolved_base,
-    xtream_choose_stream_base,
-    xtream_get_php_url,
-    xtream_player_api_url,
-)
-
-logger = logging.getLogger(__name__)
 
 from content_loader import ContentLoader
 from image_loader import ImageLoader
+from mixins import ContentLoadingMixin, DisplayMixin, ExportMixin, PlaybackMixin
 from options import OptionsDialog
-from services.export import save_m3u_content, save_stb_content
-from services.m3u import parse_m3u
-from widgets.autoplay_dialogs import (
-    EpisodeAutoPlayDialog,
-    MovieSuggestionDialog,
-    NoCategoryMoviesDialog,
-    ResumeCountdownDialog,
-    SeriesCompleteDialog,
-)
-from widgets.delegates import ChannelItemDelegate, HtmlItemDelegate
+from widgets.delegates import HtmlItemDelegate
 from widgets.menu_bar import AppMenuBar
 from widgets.sidebar import Sidebar
 from widgets.top_bar import TopBar
+from workers import SetProviderThread
 
-# --- Roman numeral helpers (conservative matching) ---
-_ROMAN_RE = re.compile(r"^M{0,3}(CM|CD|D?C{0,3})(XC|XL|L?X{0,3})(IX|IV|V?I{0,3})$")
+logger = logging.getLogger(__name__)
 
 
-def roman_to_int(token: str) -> Optional[int]:
-    if not token or token != token.upper():
-        return None
-    if not _ROMAN_RE.match(token):
-        return None
-    values = {"I": 1, "V": 5, "X": 10, "L": 50, "C": 100, "D": 500, "M": 1000}
-    total = 0
-    i = 0
-    while i < len(token):
-        if i + 1 < len(token) and values[token[i]] < values[token[i + 1]]:
-            total += values[token[i + 1]] - values[token[i]]
-            i += 2
-        else:
-            total += values[token[i]]
-            i += 1
-    return total if 1 <= total <= 3999 else None
-
-
-def find_roman_token(text: str) -> Optional[int]:
-    if not isinstance(text, str) or not text:
-        return None
-    # 1) Context keywords (Season/Episode/Part) followed by Roman numerals
-    m = re.search(r"(?i)\b(season|episode|part)\s+([MDCLXVI]+)\b", text)
-    if m:
-        token = m.group(2)
-        val = roman_to_int(token)
-        if val is not None:
-            return val
-    # 2) Trailing Roman numerals at end (e.g., "Rocky II")
-    m = re.search(r"\b([MDCLXVI]+)\)?\s*$", text)
-    if m:
-        token = m.group(1)
-        val = roman_to_int(token)
-        # Be conservative: avoid accepting single 'I' at end (too ambiguous)
-        if val is not None and val >= 2:
-            return val
-    return None
-
-
-class M3ULoaderWorker(QObject):
-    finished = Signal(object)
-    error = Signal(str)
-
-    def __init__(self, url: str, verify_ssl: bool = True, prefer_https: bool = False):
-        super().__init__()
-        self.url = url
-        self.verify_ssl = verify_ssl
-        self.prefer_https = prefer_https
-
-    def run(self):
-        try:
-            candidate_urls = []
-            if self.prefer_https and self.url.startswith("http://"):
-                candidate_urls.append("https://" + self.url[len("http://") :])
-            candidate_urls.append(self.url)
-
-            last_exc = None
-            for u in candidate_urls:
-                try:
-                    response = requests.get(u, timeout=10, verify=self.verify_ssl)
-                    response.raise_for_status()
-                    self.finished.emit({"content": response.text})
-                    return
-                except requests.RequestException as e:
-                    last_exc = e
-                    continue
-            # If we got here, all attempts failed
-            raise last_exc or Exception("Failed to load M3U")
-        except requests.RequestException as e:
-            self.error.emit(str(e))
-
-
-class XtreamAuthWorker(QObject):
-    finished = Signal(object)
-    error = Signal(str)
-
-    def __init__(
-        self,
-        base_url: str,
-        username: str,
-        password: str,
-        verify_ssl: bool = True,
-        prefer_https: bool = False,
-    ):
-        super().__init__()
-        self.base_url = base_url
-        self.username = username
-        self.password = password
-        self.verify_ssl = verify_ssl
-        self.prefer_https = prefer_https
-
-    def run(self):
-        try:
-            # First authenticate to resolve the correct domain/ports
-            auth_url = xtream_player_api_url(self.base_url, self.username, self.password)
-            resp = requests.get(auth_url, timeout=10, verify=self.verify_ssl)
-            resp.raise_for_status()
-            body = resp.json()
-            server_info = body.get("server_info", {}) if isinstance(body, dict) else {}
-            allowed: List[str] = []
-            user_info = body.get("user_info", {}) if isinstance(body, dict) else {}
-            try:
-                allowed = user_info.get("allowed_output_formats", []) or []
-            except Exception:
-                allowed = []
-
-            resolved_base = xtream_choose_resolved_base(
-                server_info, self.base_url, prefer_https=self.prefer_https
-            )
-            if not resolved_base:
-                # Fall back to given base
-                url_obj = URLObject(self.base_url)
-                resolved_base = f"{url_obj.scheme or 'http'}://{url_obj.netloc or url_obj}".rstrip(
-                    "/"
-                )
-
-            self.finished.emit(
-                {
-                    "resolved_base": resolved_base,
-                    "allowed_output_formats": allowed,
-                    "server_info": server_info,
-                }
-            )
-        except Exception as e:
-            self.error.emit(str(e))
-
-
-class XtreamLoaderWorker(QObject):
-    """Worker for loading Xtream content (Live/VOD/Series) using Player API v2."""
-
-    finished = Signal(object)
-    error = Signal(str)
-
-    def __init__(
-        self,
-        base_url: str,
-        username: str,
-        password: str,
-        content_type: str = "itv",
-        verify_ssl: bool = True,
-        prefer_https: bool = False,
-    ):
-        super().__init__()
-        self.base_url = base_url
-        self.username = username
-        self.password = password
-        self.content_type = content_type  # "itv", "vod", or "series"
-        self.verify_ssl = verify_ssl
-        self.prefer_https = prefer_https
-
-    def run(self):
-        # Use a single Session so connections are reused and properly
-        # closed when the worker finishes (avoids leaking connections
-        # against providers with strict connection limits).
-        session = requests.Session()
-        try:
-            # Step 1: Authenticate and resolve proper base + formats
-            auth_url = xtream_player_api_url(self.base_url, self.username, self.password)
-            auth_resp = session.get(auth_url, timeout=10, verify=self.verify_ssl)
-            auth_resp.raise_for_status()
-            auth_body = auth_resp.json() if auth_resp.content else {}
-            server_info = auth_body.get("server_info", {}) if isinstance(auth_body, dict) else {}
-            user_info = auth_body.get("user_info", {}) if isinstance(auth_body, dict) else {}
-            allowed_formats = user_info.get("allowed_output_formats", []) or []
-
-            resolved_base = xtream_choose_resolved_base(
-                server_info, self.base_url, prefer_https=self.prefer_https
-            )
-            if not resolved_base:
-                url_obj = URLObject(self.base_url)
-                resolved_base = f"{url_obj.scheme or 'http'}://{url_obj.netloc or url_obj}".rstrip(
-                    "/"
-                )
-
-            stream_base = xtream_choose_stream_base(server_info) or resolved_base
-
-            # Build candidate ext/base combinations
-            exts_pref = []
-            if not allowed_formats:
-                exts_pref = ["ts", "m3u8"]
-            else:
-                # Prefer TS if present, then m3u8
-                if "ts" in allowed_formats:
-                    exts_pref.append("ts")
-                if "m3u8" in allowed_formats:
-                    exts_pref.append("m3u8")
-                # Ensure at least one
-                if not exts_pref:
-                    exts_pref = ["ts"]
-
-            bases_pref = []
-            # Per API doc: prefer HTTPS when available for API & player requests
-            if resolved_base:
-                bases_pref.append(resolved_base)
-            if stream_base and stream_base not in bases_pref:
-                bases_pref.append(stream_base)
-
-            # Step 2: Determine API actions based on content type
-            if self.content_type == "itv":
-                cat_action = "get_live_categories"
-                streams_action = "get_live_streams"
-                url_prefix = "live"
-            elif self.content_type == "vod":
-                cat_action = "get_vod_categories"
-                streams_action = "get_vod_streams"
-                url_prefix = "movie"
-            elif self.content_type == "series":
-                cat_action = "get_series_categories"
-                streams_action = "get_series"
-                url_prefix = "series"
-            else:
-                raise ValueError(f"Unknown content type: {self.content_type}")
-
-            # Step 3: Fetch categories
-            cat_url = xtream_player_api_url(
-                resolved_base, self.username, self.password, action=cat_action
-            )
-            categories = []
-            categories_map = {}
-            try:
-                cat_resp = session.get(cat_url, timeout=10, verify=self.verify_ssl)
-                if cat_resp.ok and cat_resp.content:
-                    cats_data = cat_resp.json() or []
-                    for c in cats_data:
-                        cid = str(c.get("category_id"))
-                        cname = c.get("category_name") or "Unknown"
-                        categories_map[cid] = cname
-                        categories.append({"id": cid, "title": cname})
-            except Exception as e:
-                logger.warning(f"Failed to fetch categories: {e}")
-
-            # Add "All" category
-            categories.insert(0, {"id": "*", "title": "All"})
-
-            # Step 4: Fetch all content
-            streams_url = xtream_player_api_url(
-                resolved_base, self.username, self.password, action=streams_action
-            )
-            streams_resp = session.get(streams_url, timeout=15, verify=self.verify_ssl)
-            streams_resp.raise_for_status()
-            streams = streams_resp.json() or []
-
-            # Step 5: Probe a working combination for live/vod (not needed for series metadata)
-            pick_base = bases_pref[0] if bases_pref else resolved_base
-            pick_ext = exts_pref[0]
-
-            if self.content_type in ("itv", "vod"):
-                sample_id = None
-                for s in streams:
-                    sid = s.get("stream_id")
-                    if sid:
-                        sample_id = sid
-                        break
-
-                if sample_id:
-
-                    def looks_like_m3u8(rbytes, ctype):
-                        if ctype:
-                            ctype = ctype.lower()
-                            if (
-                                "application/vnd.apple.mpegurl" in ctype
-                                or "application/x-mpegurl" in ctype
-                            ):
-                                return True
-                            if "text/plain" in ctype and rbytes.startswith(b"#EXTM3U"):
-                                return True
-                        return rbytes.startswith(b"#EXTM3U")
-
-                    def looks_like_ts(rbytes, ctype):
-                        # First byte of TS packet is 0x47 (sync byte) every 188 bytes
-                        if not rbytes:
-                            return False
-                        if rbytes[0:1] == b"\x47":
-                            return True
-                        if ctype:
-                            ctype = ctype.lower()
-                            if "video/" in ctype or "application/octet-stream" in ctype:
-                                return True
-                        return False
-
-                    for b in bases_pref:
-                        worked = False
-                        # Prefer TS over HLS for live streams.
-                        # The HLS adaptive demuxer struggles with mid-stream
-                        # format changes (e.g. intro clip transitions), while
-                        # VLC's TS demuxer handles them inline more gracefully.
-                        ext_order = [ext for ext in ("ts", "m3u8") if ext in exts_pref]
-                        for ext in ext_order:
-                            test_url = f"{b}/{url_prefix}/{self.username}/{self.password}/{sample_id}.{ext}"
-                            try:
-                                headers = {
-                                    "User-Agent": "VLC/3.0.20",
-                                }
-                                # Fetch a small range sufficient to identify TS or M3U
-                                read_size = 188 if ext == "ts" else 1024
-                                if ext == "ts":
-                                    headers["Range"] = "bytes=0-187"
-                                else:
-                                    headers["Range"] = "bytes=0-1023"
-                                # Use stream=True to avoid downloading entire
-                                # live streams when the server ignores Range.
-                                r = session.get(
-                                    test_url,
-                                    headers=headers,
-                                    timeout=6,
-                                    allow_redirects=True,
-                                    verify=self.verify_ssl,
-                                    stream=True,
-                                )
-                                # Read only the bytes we need, then close
-                                probe_bytes = r.raw.read(read_size)
-                                ctype = r.headers.get("Content-Type", "")
-                                clen = r.headers.get("Content-Length", "1")
-                                r.close()
-                                if (
-                                    r.status_code in (200, 206)
-                                    and probe_bytes
-                                    and len(probe_bytes) > 0
-                                    and clen != "0"
-                                ):
-                                    ok = (
-                                        looks_like_m3u8(probe_bytes, ctype)
-                                        if ext == "m3u8"
-                                        else looks_like_ts(probe_bytes, ctype)
-                                    )
-                                    if ok:
-                                        pick_base, pick_ext = b, ext
-                                        worked = True
-                                        break
-                            except Exception:
-                                pass
-                        if worked:
-                            break
-
-                try:
-                    logger.info(
-                        "Xtream stream probe picked base=%s ext=%s (candidates bases=%s, exts=%s)",
-                        pick_base,
-                        pick_ext,
-                        bases_pref,
-                        exts_pref,
-                    )
-                except Exception:
-                    pass
-
-            # Step 6: Build items list with category mapping
-            items: List[Dict] = []
-            sorted_channels: Dict[str, List[int]] = {}
-
-            for s in streams:
-                try:
-                    stream_id = s.get("stream_id") or s.get("series_id")
-                    if not stream_id:
-                        continue
-
-                    num = s.get("num") or len(items) + 1
-                    name = s.get("name") or f"Stream {stream_id}"
-                    logo = s.get("stream_icon") or s.get("cover") or ""
-
-                    # Get category ID
-                    if isinstance(s.get("category_ids"), list) and s.get("category_ids"):
-                        cid = str(s.get("category_ids")[0])
-                    else:
-                        cid = str(s.get("category_id") or "None")
-
-                    # Build URL based on content type
-                    if self.content_type == "series":
-                        # Series don't have direct playback URLs; store series_id for later fetching
-                        cmd = ""  # Will be populated when episodes are fetched
-                    elif self.content_type == "vod":
-                        # For VOD, use container_extension from stream data
-                        container_ext = s.get("container_extension") or pick_ext
-                        cmd = f"{pick_base}/{url_prefix}/{self.username}/{self.password}/{stream_id}.{container_ext}"
-                    else:
-                        # For live streams (itv), use probed extension
-                        cmd = f"{pick_base}/{url_prefix}/{self.username}/{self.password}/{stream_id}.{pick_ext}"
-
-                    item = {
-                        "id": stream_id,
-                        "number": str(num),
-                        "name": name,
-                        "logo": logo,
-                        "tv_genre_id": cid,  # Use STB-compatible field name
-                        "cmd": cmd,
-                    }
-
-                    if self.content_type == "itv":
-                        item["xmltv_id"] = s.get("epg_channel_id") or ""
-                    elif self.content_type == "vod":
-                        item["director"] = s.get("director") or ""
-                        item["description"] = s.get("plot") or ""
-                        item["rating"] = s.get("rating") or ""
-                        item["year"] = s.get("releasedate") or ""
-                        # Store container extension for reference
-                        item["container_extension"] = s.get("container_extension") or pick_ext
-                    elif self.content_type == "series":
-                        item["plot"] = s.get("plot") or ""
-                        item["rating"] = s.get("rating") or ""
-                        item["year"] = s.get("year") or ""
-
-                    items.append(item)
-
-                    # Build sorted_channels mapping
-                    if cid not in sorted_channels:
-                        sorted_channels[cid] = []
-                    sorted_channels[cid].append(len(items) - 1)
-
-                except Exception as e:
-                    logger.warning(f"Failed to process stream: {e}")
-                    continue
-
-            # Sort channels within each category
-            for cat_id in sorted_channels:
-                sorted_channels[cat_id].sort(
-                    key=lambda x: int(items[x]["number"]) if items[x]["number"].isdigit() else 0
-                )
-
-            # Add "None" category if there are uncategorized items
-            if "None" in sorted_channels:
-                categories.append({"id": "None", "title": "Unknown Category"})
-
-            result = {
-                "categories": categories,
-                "contents": items,
-                "sorted_channels": sorted_channels,
-                "resolved_base": resolved_base,
-                "stream_base": pick_base,
-                "stream_ext": pick_ext,
-            }
-
-            self.finished.emit(result)
-
-        except Exception as e:
-            self.error.emit(str(e))
-        finally:
-            session.close()
-
-
-class XtreamSeriesInfoWorker(QObject):
-    """Worker for loading Xtream series info (seasons and episodes) using Player API v2."""
-
-    finished = Signal(object)
-    error = Signal(str)
-
-    def __init__(
-        self,
-        base_url: str,
-        username: str,
-        password: str,
-        series_id: str,
-        resolved_base: str = "",
-        verify_ssl: bool = True,
-        prefer_https: bool = False,
-    ):
-        super().__init__()
-        self.base_url = base_url
-        self.username = username
-        self.password = password
-        self.series_id = series_id
-        self.resolved_base = resolved_base
-        self.verify_ssl = verify_ssl
-        self.prefer_https = prefer_https
-
-    def run(self):
-        try:
-            # Use resolved_base if provided, otherwise determine from base_url
-            if not self.resolved_base:
-                auth_url = xtream_player_api_url(self.base_url, self.username, self.password)
-                auth_resp = requests.get(auth_url, timeout=10, verify=self.verify_ssl)
-                auth_resp.raise_for_status()
-                auth_body = auth_resp.json() if auth_resp.content else {}
-                server_info = (
-                    auth_body.get("server_info", {}) if isinstance(auth_body, dict) else {}
-                )
-                self.resolved_base = xtream_choose_resolved_base(
-                    server_info, self.base_url, prefer_https=self.prefer_https
-                )
-
-            if not self.resolved_base:
-                url_obj = URLObject(self.base_url)
-                self.resolved_base = (
-                    f"{url_obj.scheme or 'http'}://{url_obj.netloc or url_obj}".rstrip("/")
-                )
-
-            # Fetch series info
-            info_url = xtream_player_api_url(
-                self.resolved_base,
-                self.username,
-                self.password,
-                action="get_series_info",
-                extra={"series_id": self.series_id},
-            )
-            info_resp = requests.get(info_url, timeout=15, verify=self.verify_ssl)
-            info_resp.raise_for_status()
-            series_data = info_resp.json() or {}
-
-            # Extract seasons and episodes with safety checks
-            if not isinstance(series_data, dict):
-                raise ValueError(f"Invalid series_data type: {type(series_data)}")
-
-            episodes_dict = series_data.get("episodes", {})
-            if not isinstance(episodes_dict, dict):
-                episodes_dict = {}
-            seasons_info_raw = series_data.get("seasons", [])
-            series_info = series_data.get("info", {})
-
-            # Convert seasons list to dict for easy lookup (some providers use list, some use dict)
-            seasons_info = {}
-            if isinstance(seasons_info_raw, list):
-                # List format: convert to dict by season number
-                for season_item in seasons_info_raw:
-                    if isinstance(season_item, dict):
-                        season_num = str(
-                            season_item.get("season_number", season_item.get("id", ""))
-                        )
-                        if season_num:
-                            seasons_info[season_num] = season_item
-            elif isinstance(seasons_info_raw, dict):
-                # Already in dict format
-                seasons_info = seasons_info_raw
-
-            # Build seasons list
-            seasons = []
-            if episodes_dict:
-                for season_num in sorted(
-                    episodes_dict.keys(), key=lambda x: int(x) if x.isdigit() else 0
-                ):
-                    season_episodes = episodes_dict.get(season_num, [])
-                    if not isinstance(season_episodes, list):
-                        continue
-
-                    season_data = seasons_info.get(season_num, {})
-                    if not isinstance(season_data, dict):
-                        season_data = {}
-
-                    season_item = {
-                        "id": season_num,
-                        "name": season_data.get("name") or f"Season {season_num}",
-                        "cover": season_data.get("cover_big") or season_data.get("cover") or "",
-                        "overview": season_data.get("overview") or "",
-                        "air_date": season_data.get("air_date") or "",
-                        "episode_count": str(len(season_episodes)),
-                        "episodes": season_episodes,  # Store episodes with season
-                    }
-                    seasons.append(season_item)
-
-            # Explicitly convert to plain Python types
-            result = {
-                "seasons": list(seasons),  # Ensure it's a plain list
-                "series_info": dict(series_info) if series_info else {},
-                "resolved_base": str(self.resolved_base) if self.resolved_base else "",
-            }
-
-            # Emit result to UI
-            self.finished.emit(result)
-
-        except Exception as e:
-            logger.error(f"Worker error: {e}")
-            try:
-                self.error.emit(str(e))
-            except Exception as emit_err:
-                logger.error(f"Error emitting error: {emit_err}")
-
-
-class STBCategoriesWorker(QObject):
-    finished = Signal(object)
-    error = Signal(str)
-
-    def __init__(
-        self,
-        base_url: str,
-        headers: dict,
-        content_type: str = "itv",
-        verify_ssl: bool = True,
-        prefer_https: bool = False,
-    ):
-        super().__init__()
-        self.base_url = base_url
-        self.headers = headers
-        self.content_type = content_type
-        self.verify_ssl = verify_ssl
-        self.prefer_https = prefer_https
-
-    def run(self):
-        try:
-            base = base_from_url(self.base_url)
-            if self.prefer_https and base.startswith("http://"):
-                base = "https://" + base[len("http://") :]
-
-            # Use correct action based on content type
-            action = "get_genres" if self.content_type == "itv" else "get_categories"
-            fetchurl = stb_request_url(base, self.content_type, action)
-            resp = requests.get(fetchurl, headers=self.headers, timeout=10, verify=self.verify_ssl)
-            resp.raise_for_status()
-            categories = resp.json()["js"]
-
-            # Only fetch all channels for itv type
-            all_channels = []
-            if self.content_type == "itv":
-                fetchurl = stb_request_url(base, "itv", "get_all_channels")
-                resp = requests.get(
-                    fetchurl, headers=self.headers, timeout=10, verify=self.verify_ssl
-                )
-                resp.raise_for_status()
-                all_channels = resp.json()["js"]["data"]
-
-            self.finished.emit({"categories": categories, "all_channels": all_channels})
-        except Exception as e:
-            self.error.emit(str(e))
-
-
-class LinkCreatorWorker(QObject):
-    finished = Signal(object)
-    error = Signal(str)
-
-    def __init__(
-        self,
-        base_url: str,
-        headers: dict,
-        content_type: str,
-        cmd: str,
-        is_episode: bool = False,
-        series_param: Optional[str] = None,
-        verify_ssl: bool = True,
-        prefer_https: bool = False,
-    ):
-        super().__init__()
-        self.base_url = base_url
-        self.headers = headers
-        self.content_type = content_type
-        self.cmd = cmd
-        self.is_episode = is_episode
-        self.series_param = series_param
-        self.verify_ssl = verify_ssl
-        self.prefer_https = prefer_https
-
-    def run(self):
-        try:
-            url = URLObject(self.base_url)
-            scheme = url.scheme
-            if self.prefer_https and scheme == "http":
-                scheme = "https"
-            base = f"{scheme}://{url.netloc}"
-            if self.is_episode:
-                fetchurl = (
-                    f"{base}/server/load.php?type={'vod' if self.content_type == 'series' else self.content_type}&action=create_link"
-                    f"&cmd={url_quote(self.cmd)}&series={self.series_param}&JsHttpRequest=1-xml"
-                )
-            else:
-                fetchurl = (
-                    f"{base}/server/load.php?type={self.content_type}&action=create_link"
-                    f"&cmd={url_quote(self.cmd)}&JsHttpRequest=1-xml"
-                )
-            response = requests.get(
-                fetchurl, headers=self.headers, timeout=10, verify=self.verify_ssl
-            )
-            response.raise_for_status()
-            result = response.json()
-            link = result["js"]["cmd"].split(" ")[-1]
-            self.finished.emit({"link": link})
-        except Exception as e:
-            self.error.emit(str(e))
-
-
-class CategoryTreeWidgetItem(QTreeWidgetItem):
-    # sort to always have value "All" first and "Unknown Category" last
-    def __lt__(self, other):
-        if not isinstance(other, CategoryTreeWidgetItem):
-            return super(CategoryTreeWidgetItem, self).__lt__(other)
-
-        sort_column = self.treeWidget().sortColumn()
-        t1 = self.text(sort_column)
-        t2 = other.text(sort_column)
-        if t1 == "All":
-            return True
-        if t2 == "All":
-            return False
-        if t1 == "Unknown Category":
-            return False
-        if t2 == "Unknown Category":
-            return True
-        return t1 < t2
-
-
-class ChannelTreeWidgetItem(QTreeWidgetItem):
-    # Modify the sorting by Channel Number to used integer and not string (1 < 10, but "1" may not be < "10")
-    # Modify the sorting by Program Progress to read the progress in item data
-    def __lt__(self, other):
-        if not isinstance(other, ChannelTreeWidgetItem):
-            return super(ChannelTreeWidgetItem, self).__lt__(other)
-
-        sort_column = self.treeWidget().sortColumn()
-        if sort_column == 0:  # Channel number
-            return int(self.text(sort_column)) < int(other.text(sort_column))
-        elif sort_column == 2:  # EPG Program progress
-            p1 = self.data(sort_column, Qt.UserRole)
-            if p1 is None:
-                return False
-            p2 = other.data(sort_column, Qt.UserRole)
-            if p2 is None:
-                return True
-            return self.data(sort_column, Qt.UserRole) < other.data(sort_column, Qt.UserRole)
-        elif sort_column == 3:  # EPG Program name
-            return self.data(sort_column, Qt.UserRole) < other.data(sort_column, Qt.UserRole)
-
-        return self.text(sort_column) < other.text(sort_column)
-
-
-class NumberedTreeWidgetItem(QTreeWidgetItem):
-    # Modify the sorting by Number to use integer and not string (1 < 10, but "1" may not be < "10")
-    def __lt__(self, other):
-        if not isinstance(other, NumberedTreeWidgetItem):
-            return super(NumberedTreeWidgetItem, self).__lt__(other)
-
-        # Safety check: ensure widget is available
-        try:
-            widget = self.treeWidget()
-            if not widget:
-                return super(NumberedTreeWidgetItem, self).__lt__(other)
-
-            sort_column = widget.sortColumn()
-            if sort_column == 0:  # Number column (channel/season/episode number)
-                # Safely convert to int, fallback to string comparison
-                try:
-                    return int(self.text(sort_column)) < int(other.text(sort_column))
-                except (ValueError, TypeError):
-                    return self.text(sort_column) < other.text(sort_column)
-            return self.text(sort_column) < other.text(sort_column)
-        except (RuntimeError, AttributeError):
-            # Widget deleted or in invalid state
-            return super(NumberedTreeWidgetItem, self).__lt__(other)
-
-
-## Delegates moved to widgets/delegates.py
-
-
-class SetProviderThread(QThread):
-    progress = Signal(str)
-
-    def __init__(self, provider_manager, epg_manager, force_epg_refresh: bool = False):
-        super().__init__()
-        self.provider_manager = provider_manager
-        self.epg_manager = epg_manager
-        self.force_epg_refresh = force_epg_refresh
-
-    def run(self):
-        try:
-            self.provider_manager.set_current_provider(self.progress)
-            if self.force_epg_refresh:
-                try:
-                    self.progress.emit("Refreshing EPG…")
-                except Exception:
-                    pass
-                self.epg_manager.force_refresh_current_epg()
-            else:
-                self.epg_manager.set_current_epg()
-        except Exception as e:
-            logger.warning(f"Error in initializing provider: {e}")
-
-
-class ChannelList(QMainWindow):
+class ChannelList(
+    PlaybackMixin,
+    ContentLoadingMixin,
+    ExportMixin,
+    DisplayMixin,
+    QMainWindow,
+):
     def __init__(self, app, player, config_manager, provider_manager, image_manager, epg_manager):
         super().__init__()
         self.app = app
@@ -966,6 +181,10 @@ class ChannelList(QMainWindow):
         # Keep references to background jobs (threads/workers)
         self._bg_jobs = []
 
+    # ------------------------------------------------------------------
+    # Lifecycle
+    # ------------------------------------------------------------------
+
     def closeEvent(self, event):
         # Stop and delete timer
         if self.refresh_on_air_timer.isActive():
@@ -978,6 +197,10 @@ class ChannelList(QMainWindow):
         self.epg_manager.save_index()
         self.config_manager.save_window_settings(self, "channel_list")
         event.accept()
+
+    # ------------------------------------------------------------------
+    # EPG refresh
+    # ------------------------------------------------------------------
 
     def refresh_on_air(self):
         for i in range(self.content_list.topLevelItemCount()):
@@ -1046,6 +269,10 @@ class ChannelList(QMainWindow):
                     item.setData(3, Qt.UserRole, "")
 
         self.content_list.viewport().update()
+
+    # ------------------------------------------------------------------
+    # Provider management
+    # ------------------------------------------------------------------
 
     def set_provider(self, force_update=False):
         self.lock_ui_before_loading()
@@ -1145,6 +372,58 @@ class ChannelList(QMainWindow):
         force_update = getattr(self, "_set_provider_force_update", False)
         self.set_provider_finished(force_update)
 
+    def populate_provider_combo(self):
+        """Populate the provider dropdown with available providers."""
+        # Block signals to prevent triggering change during population
+        was_blocked = self.provider_combo.blockSignals(True)
+
+        try:
+            self.provider_combo.clear()
+
+            for provider in self.provider_manager.providers:
+                self.provider_combo.addItem(provider["name"])
+
+            # Set current provider
+            current_name = self.config_manager.selected_provider_name
+            index = self.provider_combo.findText(current_name)
+            if index >= 0:
+                self.provider_combo.setCurrentIndex(index)
+        finally:
+            # Restore previous signal blocking state
+            self.provider_combo.blockSignals(was_blocked)
+
+        # Also populate sidebar and menu bar if they exist
+        provider_names = [
+            self.provider_combo.itemText(i) for i in range(self.provider_combo.count())
+        ]
+        current_name = self.config_manager.selected_provider_name
+        if hasattr(self, "sidebar"):
+            self.sidebar.set_providers(provider_names)
+            if current_name:
+                self.sidebar.select_provider(current_name)
+        if hasattr(self, "app_menu"):
+            self.app_menu.set_providers(provider_names, current_name)
+
+    def on_provider_changed(self, provider_name):
+        """Handle provider selection change from combo box."""
+        if not provider_name:
+            return
+
+        # Check if this is actually a change
+        if provider_name == self.config_manager.selected_provider_name:
+            return
+
+        # Update config
+        self.config_manager.selected_provider_name = provider_name
+        self.config_manager.save_config()
+
+        # Reload provider (use QTimer to ensure we're in the main thread)
+        QTimer.singleShot(0, lambda: self.set_provider())
+
+    # ------------------------------------------------------------------
+    # Splitter management
+    # ------------------------------------------------------------------
+
     def update_splitter_ratio(self, pos, index):
         sizes = self.splitter.sizes()
         total_size = sizes[0] + sizes[1]
@@ -1156,6 +435,10 @@ class ChannelList(QMainWindow):
         total_size = sizes[0] + sizes[1]
         if total_size:
             self.splitter_content_info_ratio = sizes[0] / total_size
+
+    # ------------------------------------------------------------------
+    # UI setup
+    # ------------------------------------------------------------------
 
     def create_top_bar(self):
         self.top_bar = TopBar(self.container_widget)
@@ -1251,91 +534,113 @@ class ChannelList(QMainWindow):
         act_forward.triggered.connect(self.go_forward)
         self.addAction(act_forward)
 
-    def _is_playable_item(self, item: QTreeWidgetItem) -> bool:
-        try:
-            data = item.data(0, Qt.UserRole)
-            t = data.get("type") if isinstance(data, dict) else None
-            return t in {"m3ucontent", "channel", "movie", "episode"}
-        except Exception:
-            return False
+    def create_list_panel(self):
+        self.list_panel = QWidget(self.container_widget)
+        list_layout = QVBoxLayout(self.list_panel)
+        list_layout.setContentsMargins(0, 0, 0, 0)
 
-    def channel_surf_next(self):
-        """Move selection down by one; auto-play only if playable (not folders)."""
-        cl = self.content_list
-        count = cl.topLevelItemCount()
-        if count == 0:
-            return
-        current = cl.currentItem()
-        idx = cl.indexOfTopLevelItem(current) if current else -1
-        idx = (idx + 1) % count
-        candidate = cl.topLevelItem(idx)
-        if candidate is not None:
-            cl.setCurrentItem(candidate)
-            if self._is_playable_item(candidate):
-                self.item_activated(candidate)
+        self.content_list = QTreeWidget(self.list_panel)
+        self.content_list.setSelectionMode(QTreeWidget.SingleSelection)
+        self.content_list.setIndentation(0)
+        self.content_list.setAlternatingRowColors(True)
+        self.content_list.itemSelectionChanged.connect(self.item_selected)
+        self.content_list.itemActivated.connect(self.item_activated)
+        self.content_list.installEventFilter(self)
+        self.content_list.setContextMenuPolicy(Qt.CustomContextMenu)
+        self.content_list.customContextMenuRequested.connect(self.show_content_context_menu)
+        self.refresh_content_list_size()
 
-    def channel_surf_prev(self):
-        """Move selection up by one; auto-play only if playable (not folders)."""
-        cl = self.content_list
-        count = cl.topLevelItemCount()
-        if count == 0:
-            return
-        current = cl.currentItem()
-        idx = cl.indexOfTopLevelItem(current) if current else 0
-        idx = (idx - 1) % count
-        candidate = cl.topLevelItem(idx)
-        if candidate is not None:
-            cl.setCurrentItem(candidate)
-            if self._is_playable_item(candidate):
-                self.item_activated(candidate)
+        list_layout.addWidget(self.content_list, 1)
 
-    def populate_provider_combo(self):
-        """Populate the provider dropdown with available providers."""
-        # Block signals to prevent triggering change during population
-        was_blocked = self.provider_combo.blockSignals(True)
+        self.progress_bar = QProgressBar(self)
+        self.progress_bar.setRange(0, 100)
+        self.progress_bar.setValue(0)
+        self.progress_bar.setVisible(False)
+        list_layout.addWidget(self.progress_bar)
 
-        try:
-            self.provider_combo.clear()
+        self.cancel_button = QPushButton("Cancel")
+        self.cancel_button.clicked.connect(self.cancel_loading)
+        self.cancel_button.setVisible(False)
+        list_layout.addWidget(self.cancel_button)
 
-            for provider in self.provider_manager.providers:
-                self.provider_combo.addItem(provider["name"])
+    def create_content_info_panel(self):
+        self.content_info_panel = QWidget(self.container_widget)
+        self.content_info_layout = QVBoxLayout(self.content_info_panel)
+        self.content_info_panel.setVisible(False)
 
-            # Set current provider
-            current_name = self.config_manager.selected_provider_name
-            index = self.provider_combo.findText(current_name)
-            if index >= 0:
-                self.provider_combo.setCurrentIndex(index)
-        finally:
-            # Restore previous signal blocking state
-            self.provider_combo.blockSignals(was_blocked)
+    def setup_movie_tvshow_content_info(self):
+        self.clear_content_info_panel()
+        self.content_info_text = QLabel(self.content_info_panel)
+        self.content_info_text.setSizePolicy(
+            QSizePolicy.Expanding, QSizePolicy.Ignored
+        )  # Allow to reduce splitter below label minimum size
+        self.content_info_text.setAlignment(Qt.AlignLeft | Qt.AlignTop)
+        self.content_info_text.setWordWrap(True)
+        self.content_info_layout.addWidget(self.content_info_text, 1)
+        self.content_info_shown = "movie_tvshow"
 
-        # Also populate sidebar and menu bar if they exist
-        provider_names = [
-            self.provider_combo.itemText(i) for i in range(self.provider_combo.count())
-        ]
-        current_name = self.config_manager.selected_provider_name
-        if hasattr(self, "sidebar"):
-            self.sidebar.set_providers(provider_names)
-            if current_name:
-                self.sidebar.select_provider(current_name)
-        if hasattr(self, "app_menu"):
-            self.app_menu.set_providers(provider_names, current_name)
+    def setup_channel_program_content_info(self):
+        self.clear_content_info_panel()
+        self.splitter_content_info = QSplitter(Qt.Horizontal)
+        self.program_list = QListWidget()
+        self.program_list.setHorizontalScrollBarPolicy(Qt.ScrollBarAlwaysOff)
+        self.program_list.setItemDelegate(HtmlItemDelegate())
+        self.splitter_content_info.addWidget(self.program_list)
+        self.content_info_text = QLabel()
+        self.content_info_text.setAlignment(Qt.AlignLeft | Qt.AlignTop)
+        self.content_info_text.setWordWrap(True)
+        self.splitter_content_info.addWidget(self.content_info_text)
+        self.content_info_layout.addWidget(self.splitter_content_info)
+        self.splitter_content_info.setSizes(
+            [
+                int(self.content_info_panel.width() * self.splitter_content_info_ratio),
+                int(self.content_info_panel.width() * (1 - self.splitter_content_info_ratio)),
+            ]
+        )
+        self.content_info_shown = "channel"
 
-    def on_provider_changed(self, provider_name):
-        """Handle provider selection change from combo box."""
-        if not provider_name:
-            return
+        self.program_list.itemSelectionChanged.connect(self.update_channel_program)
+        self.splitter_content_info.splitterMoved.connect(self.update_splitter_content_info_ratio)
 
-        # Check if this is actually a change
-        if provider_name == self.config_manager.selected_provider_name:
-            return
+    def clear_content_info_panel(self):
+        # Clear all widgets from the content_info layout
+        for i in reversed(range(self.content_info_layout.count())):
+            widget = self.content_info_layout.itemAt(i).widget()
+            if widget is not None:
+                widget.setParent(None)
+                widget.deleteLater()
 
-        # Update config
-        self.config_manager.selected_provider_name = provider_name
-        self.config_manager.save_config()
+        # Clear the layout itself
+        while self.content_info_layout.count():
+            item = self.content_info_layout.takeAt(0)
+            if item.widget():
+                item.widget().deleteLater()
+            elif item.layout():
+                self.clear_layout(item.layout())
 
-        # Reload provider (use QTimer to ensure we're in the main thread)
-        QTimer.singleShot(0, lambda: self.set_provider())
+        # Hide the content_info panel if it is visible
+        if self.content_info_panel.isVisible():
+            self.content_info_panel.setVisible(False)
+            self.splitter.setSizes([1, 0])
+
+        self.content_info_shown = None
+        self.update_layout()
+
+    def update_layout(self):
+        # Layout margins are handled by individual widgets (sidebar, top_bar)
+        if self.content_info_panel.isVisible():
+            self.content_info_layout.setContentsMargins(8, 4, 8, 8)
+
+    @staticmethod
+    def clear_layout(layout):
+        while layout.count():
+            item = layout.takeAt(0)
+            if item.widget():
+                item.widget().deleteLater()
+
+    # ------------------------------------------------------------------
+    # Menu & signal connections
+    # ------------------------------------------------------------------
 
     def _connect_menu_actions(self):
         """Connect menu bar actions to existing handler methods."""
@@ -1533,34 +838,9 @@ class ChannelList(QMainWindow):
         self._all_provider_cache_snapshot = []
         self.top_bar.search_box.setPlaceholderText("Search content...")
 
-    def create_list_panel(self):
-        self.list_panel = QWidget(self.container_widget)
-        list_layout = QVBoxLayout(self.list_panel)
-        list_layout.setContentsMargins(0, 0, 0, 0)
-
-        self.content_list = QTreeWidget(self.list_panel)
-        self.content_list.setSelectionMode(QTreeWidget.SingleSelection)
-        self.content_list.setIndentation(0)
-        self.content_list.setAlternatingRowColors(True)
-        self.content_list.itemSelectionChanged.connect(self.item_selected)
-        self.content_list.itemActivated.connect(self.item_activated)
-        self.content_list.installEventFilter(self)
-        self.content_list.setContextMenuPolicy(Qt.CustomContextMenu)
-        self.content_list.customContextMenuRequested.connect(self.show_content_context_menu)
-        self.refresh_content_list_size()
-
-        list_layout.addWidget(self.content_list, 1)
-
-        self.progress_bar = QProgressBar(self)
-        self.progress_bar.setRange(0, 100)
-        self.progress_bar.setValue(0)
-        self.progress_bar.setVisible(False)
-        list_layout.addWidget(self.progress_bar)
-
-        self.cancel_button = QPushButton("Cancel")
-        self.cancel_button.clicked.connect(self.cancel_loading)
-        self.cancel_button.setVisible(False)
-        list_layout.addWidget(self.cancel_button)
+    # ------------------------------------------------------------------
+    # EPG display helpers
+    # ------------------------------------------------------------------
 
     def show_vodinfo(self):
         self.config_manager.show_stb_content_info = self.app_menu.show_vod_info_action.isChecked()
@@ -1638,1664 +918,9 @@ class ChannelList(QMainWindow):
             return True
         return False
 
-    def create_content_info_panel(self):
-        self.content_info_panel = QWidget(self.container_widget)
-        self.content_info_layout = QVBoxLayout(self.content_info_panel)
-        self.content_info_panel.setVisible(False)
-
-    def setup_movie_tvshow_content_info(self):
-        self.clear_content_info_panel()
-        self.content_info_text = QLabel(self.content_info_panel)
-        self.content_info_text.setSizePolicy(
-            QSizePolicy.Expanding, QSizePolicy.Ignored
-        )  # Allow to reduce splitter below label minimum size
-        self.content_info_text.setAlignment(Qt.AlignLeft | Qt.AlignTop)
-        self.content_info_text.setWordWrap(True)
-        self.content_info_layout.addWidget(self.content_info_text, 1)
-        self.content_info_shown = "movie_tvshow"
-
-    def setup_channel_program_content_info(self):
-        self.clear_content_info_panel()
-        self.splitter_content_info = QSplitter(Qt.Horizontal)
-        self.program_list = QListWidget()
-        self.program_list.setHorizontalScrollBarPolicy(Qt.ScrollBarAlwaysOff)
-        self.program_list.setItemDelegate(HtmlItemDelegate())
-        self.splitter_content_info.addWidget(self.program_list)
-        self.content_info_text = QLabel()
-        self.content_info_text.setAlignment(Qt.AlignLeft | Qt.AlignTop)
-        self.content_info_text.setWordWrap(True)
-        self.splitter_content_info.addWidget(self.content_info_text)
-        self.content_info_layout.addWidget(self.splitter_content_info)
-        self.splitter_content_info.setSizes(
-            [
-                int(self.content_info_panel.width() * self.splitter_content_info_ratio),
-                int(self.content_info_panel.width() * (1 - self.splitter_content_info_ratio)),
-            ]
-        )
-        self.content_info_shown = "channel"
-
-        self.program_list.itemSelectionChanged.connect(self.update_channel_program)
-        self.splitter_content_info.splitterMoved.connect(self.update_splitter_content_info_ratio)
-
-    def clear_content_info_panel(self):
-        # Clear all widgets from the content_info layout
-        for i in reversed(range(self.content_info_layout.count())):
-            widget = self.content_info_layout.itemAt(i).widget()
-            if widget is not None:
-                widget.setParent(None)
-                widget.deleteLater()
-
-        # Clear the layout itself
-        while self.content_info_layout.count():
-            item = self.content_info_layout.takeAt(0)
-            if item.widget():
-                item.widget().deleteLater()
-            elif item.layout():
-                self.clear_layout(item.layout())
-
-        # Hide the content_info panel if it is visible
-        if self.content_info_panel.isVisible():
-            self.content_info_panel.setVisible(False)
-            self.splitter.setSizes([1, 0])
-
-        self.content_info_shown = None
-        self.update_layout()
-
-    def update_layout(self):
-        # Layout margins are handled by individual widgets (sidebar, top_bar)
-        if self.content_info_panel.isVisible():
-            self.content_info_layout.setContentsMargins(8, 4, 8, 8)
-
-    @staticmethod
-    def clear_layout(layout):
-        while layout.count():
-            item = layout.takeAt(0)
-            if item.widget():
-                item.widget().deleteLater()
-            elif item.layout():
-                ChannelList.clear_layout(item.layout())
-        layout.deleteLater()
-
-    def switch_content_info_panel(self, item_type):
-        if item_type in ["channel", "m3ucontent"]:
-            if self.content_info_shown == "channel":
-                return
-            self.setup_channel_program_content_info()
-        else:
-            if self.content_info_shown == "movie_tvshow":
-                return
-            self.setup_movie_tvshow_content_info()
-
-        if not self.content_info_panel.isVisible():
-            self.content_info_panel.setVisible(True)
-            self.splitter.setSizes(
-                [
-                    int(self.container_widget.height() * self.splitter_ratio),
-                    int(self.container_widget.height() * (1 - self.splitter_ratio)),
-                ]
-            )
-
-    def populate_channel_programs_content_info(self, item_data):
-        try:
-            self.program_list.itemSelectionChanged.disconnect()
-        except (TypeError, RuntimeError):
-            pass
-        self.program_list.clear()
-        self.program_list.itemSelectionChanged.connect(self.update_channel_program)
-
-        # Show EPG data for the selected channel
-        # Show full EPG list for the channel (no windowing)
-        epg_data = self.epg_manager.get_programs_for_channel(item_data, max_programs=0)
-        if epg_data:
-            # Fill the program list and try to select the currently playing entry
-            now_index = None
-            try:
-                import tzlocal as _tz
-
-                _local_now = datetime.now(_tz.get_localzone())
-            except Exception:
-                _local_now = datetime.now()
-            for idx, epg_item in enumerate(epg_data):
-                # Detect structure and format row text accordingly
-                if "@start" in epg_item and "@stop" in epg_item:
-                    try:
-                        local_tz = tzlocal.get_localzone()
-                        start_raw = datetime.strptime(epg_item.get("@start"), "%Y%m%d%H%M%S %z")
-                        stop_raw = datetime.strptime(epg_item.get("@stop"), "%Y%m%d%H%M%S %z")
-                        start_dt = start_raw.astimezone(local_tz)
-                        stop_dt = stop_raw.astimezone(local_tz)
-                        start_txt = start_dt.strftime("%H:%M")
-                        stop_txt = stop_dt.strftime("%H:%M")
-                        if (
-                            now_index is None
-                            and start_raw <= _local_now.astimezone(start_raw.tzinfo) < stop_raw
-                        ):
-                            now_index = idx
-                    except Exception:
-                        start_txt, stop_txt = "", ""
-                    title_val = epg_item.get("title")
-                    title_txt = ""
-                    if isinstance(title_val, dict):
-                        title_txt = title_val.get("__text") or ""
-                    elif isinstance(title_val, list) and title_val:
-                        first = title_val[0]
-                        if isinstance(first, dict):
-                            title_txt = first.get("__text") or ""
-                    epg_text = f"<b>{start_txt}-{stop_txt}</b>&nbsp;&nbsp;{title_txt}"
-                else:
-                    # STB style: t_time fields may exist; fallback to parsed time/time_to
-                    start_txt = epg_item.get("t_time")
-                    stop_txt = epg_item.get("t_time_to")
-                    if (
-                        not (start_txt and stop_txt)
-                        and "time" in epg_item
-                        and "time_to" in epg_item
-                    ):
-                        try:
-                            start_dt = datetime.strptime(epg_item.get("time"), "%Y-%m-%d %H:%M:%S")
-                            stop_dt = datetime.strptime(
-                                epg_item.get("time_to"), "%Y-%m-%d %H:%M:%S"
-                            )
-                            start_txt = start_dt.strftime("%H:%M")
-                            stop_txt = stop_dt.strftime("%H:%M")
-                            if now_index is None and start_dt <= datetime.now() < stop_dt:
-                                now_index = idx
-                        except Exception:
-                            start_txt, stop_txt = "", ""
-                    epg_text = f"<b>{start_txt or ''}-{stop_txt or ''}</b>&nbsp;&nbsp;{epg_item.get('name', '')}"
-                item = QListWidgetItem(f"{epg_text}")
-                item.setData(Qt.UserRole, epg_item)
-                self.program_list.addItem(item)
-            # Visually highlight the currently airing program
-            if now_index is not None:
-                try:
-                    now_item = self.program_list.item(now_index)
-                    if now_item is not None:
-                        # Prefix with a play arrow for visibility
-                        try:
-                            current_text = now_item.text()
-                            if not current_text.lstrip().startswith("▶ Now"):
-                                now_item.setText(f"▶ Now  {current_text}")
-                        except Exception:
-                            pass
-                        # Light blue background tint
-                        try:
-                            now_item.setBackground(QColor(201, 107, 67, 45))
-                        except Exception:
-                            pass
-                except Exception:
-                    pass
-                self.program_list.setCurrentRow(now_index)
-            else:
-                self.program_list.setCurrentRow(0)
-        else:
-            item = QListWidgetItem("Program not available")
-            self.program_list.addItem(item)
-            xmltv_id = item_data.get("xmltv_id", "")
-            ch_name = item_data.get("name", "")
-            if xmltv_id:
-                self.content_info_text.setText(f'No EPG found for channel id "{xmltv_id}"')
-            elif ch_name:
-                self.content_info_text.setText(f'No EPG found for channel "{ch_name}"')
-            else:
-                self.content_info_text.setText("Channel without id")
-
-    def update_channel_program(self):
-        selected_items = self.program_list.selectedItems()
-        if not selected_items:
-            self.content_info_text.setText("No program selected")
-            return
-        selected_item = selected_items[0]
-        item_data = selected_item.data(Qt.UserRole)
-        if item_data:
-            # Decide formatting by epg item structure rather than global source
-            if "@start" not in item_data:
-                # Extract information from item_data
-                title = item_data.get("name", {})
-                desc = item_data.get("descr")
-                desc = desc.replace("\r\n", "<br>") if desc else ""
-                director = item_data.get("director")
-                actor = item_data.get("actor")
-                category = item_data.get("category")
-
-                # Format the content information
-                info = ""
-                if title:
-                    info += f"<b>Title:</b> {title}<br>"
-                if category:
-                    info += f"<b>Category:</b> {category}<br>"
-                if desc:
-                    info += f"<b>Description:</b> {desc}<br>"
-                if director:
-                    info += f"<b>Director:</b> {director}<br>"
-                if actor:
-                    info += f"<b>Actor:</b> {actor}<br>"
-
-                self.content_info_text.setText(info if info else "No data available")
-
-            else:
-                # Extract information from item_data
-                title = item_data.get("title", {})
-                sub_title = item_data.get("sub-title")
-                desc = item_data.get("desc")
-                credits = item_data.get("credits", {})
-                director = credits.get("director")
-                actor = credits.get("actor")
-                writer = credits.get("writer")
-                presenter = credits.get("presenter")
-                adapter = credits.get("adapter")
-                producer = credits.get("producer")
-                composer = credits.get("composer")
-                editor = credits.get("editor")
-                guest = credits.get("guest")
-                category = item_data.get("category")
-                country = item_data.get("country")
-                episode_num = item_data.get("episode-num")
-                rating = item_data.get("rating", {}).get("value")
-
-                # Format the content information
-                info = ""
-                if title:
-                    info += f"<b>Title:</b> {title.get('__text')}<br>"
-                if sub_title:
-                    info += f"<b>Sub-title:</b> {sub_title.get('__text')}<br>"
-                if episode_num:
-                    info += f"<b>Episode Number:</b> {episode_num.get('__text')}<br>"
-                if category:
-                    if isinstance(category, dict):
-                        info += f"<b>Category:</b> {category.get('__text')}<br>"
-                    elif isinstance(category, list):
-                        info += (
-                            f"<b>Category:</b> {', '.join([c.get('__text') for c in category])}<br>"
-                        )
-                if rating:
-                    info += f"<b>Rating:</b> {rating.get('__text')}<br>"
-                if desc:
-                    info += f"<b>Description:</b> {desc.get('__text')}<br>"
-                if credits:
-                    if director:
-                        if isinstance(director, dict):
-                            info += f"<b>Director:</b> {director.get('__text')}<br>"
-                        elif isinstance(director, list):
-                            info += f"<b>Director:</b> {', '.join([c.get('__text') for c in director])}<br>"
-                    if actor:
-                        if isinstance(actor, dict):
-                            info += f"<b>Actor:</b> {actor.get('__text')}<br>"
-                        elif isinstance(actor, list):
-                            info += (
-                                f"<b>Actor:</b> {', '.join([c.get('__text') for c in actor])}<br>"
-                            )
-                    if guest:
-                        if isinstance(guest, dict):
-                            info += f"<b>Guest:</b> {guest.get('__text')}<br>"
-                        elif isinstance(guest, list):
-                            info += (
-                                f"<b>Guest:</b> {', '.join([c.get('__text') for c in guest])}<br>"
-                            )
-                    if writer:
-                        if isinstance(writer, dict):
-                            info += f"<b>Writer:</b> {writer.get('__text')}<br>"
-                        elif isinstance(writer, list):
-                            info += (
-                                f"<b>Writer:</b> {', '.join([c.get('__text') for c in writer])}<br>"
-                            )
-                    if presenter:
-                        if isinstance(presenter, dict):
-                            info += f"<b>Presenter:</b> {presenter.get('__text')}<br>"
-                        elif isinstance(presenter, list):
-                            info += f"<b>Presenter:</b> {', '.join([c.get('__text') for c in presenter])}<br>"
-                    if adapter:
-                        if isinstance(adapter, dict):
-                            info += f"<b>Adapter:</b> {adapter.get('__text')}<br>"
-                        elif isinstance(adapter, list):
-                            info += f"<b>Adapter:</b> {', '.join([c.get('__text') for c in adapter])}<br>"
-                    if producer:
-                        if isinstance(producer, dict):
-                            info += f"<b>Producer:</b> {producer.get('__text')}<br>"
-                        elif isinstance(producer, list):
-                            info += f"<b>Producer:</b> {', '.join([c.get('__text') for c in producer])}<br>"
-                    if composer:
-                        if isinstance(composer, dict):
-                            info += f"<b>Composer:</b> {composer.get('__text')}<br>"
-                        elif isinstance(composer, list):
-                            info += f"<b>Composer:</b> {', '.join([c.get('__text') for c in composer])}<br>"
-                    if editor:
-                        if isinstance(editor, dict):
-                            info += f"<b>Editor:</b> {editor.get('__text')}<br>"
-                        elif isinstance(editor, list):
-                            info += (
-                                f"<b>Editor:</b> {', '.join([c.get('__text') for c in editor])}<br>"
-                            )
-                if country:
-                    info += f"<b>Country:</b> {country.get('__text')}<br>"
-
-                self.content_info_text.setText(info if info else "No data available")
-
-                # Load poster image if available
-                icon_url = item_data.get("icon", {}).get("@src")
-                if icon_url:
-                    self.lock_ui_before_loading()
-                    if self.image_loader and self.image_loader.isRunning():
-                        self.image_loader.wait()
-                    self.image_loader = ImageLoader(
-                        [
-                            icon_url,
-                        ],
-                        self.image_manager,
-                        iconified=False,
-                        verify_ssl=self.config_manager.ssl_verify,
-                    )
-                    self.image_loader.progress_updated.connect(self.update_poster)
-                    self.image_loader.finished.connect(self.image_loader_finished)
-                    self.image_loader.start()
-                    self.cancel_button.setText("Cancel fetching poster...")
-        else:
-            self.content_info_text.setText("No data available")
-
-    def populate_movie_tvshow_content_info(self, item_data):
-        provider_type = self.provider_manager.current_provider.get("type", "").upper()
-
-        # Common labels; not all keys will be present for all providers
-        stb_labels = {
-            "name": "Title",
-            "rating_imdb": "Rating",
-            "year": "Year",
-            "genres_str": "Genre",
-            "length": "Length",
-            "director": "Director",
-            "actors": "Actors",
-            "description": "Summary",
-        }
-        xtream_labels = {
-            "name": "Title",
-            "rating": "Rating",
-            "year": "Year",
-            "genre": "Genre",
-            "director": "Director",
-            "actors": "Actors",
-            "description": "Summary",
-            "plot": "Summary",
-        }
-
-        labels = stb_labels if provider_type == "STB" else xtream_labels
-
-        info = ""
-        for key, label in labels.items():
-            if key in item_data:
-                value = item_data.get(key)
-                if not value:
-                    continue
-                # Normalize dict/list values from some APIs
-                if isinstance(value, dict) and "__text" in value:
-                    value = value.get("__text")
-                if isinstance(value, list):
-                    value = ", ".join(str(v) for v in value if v)
-                if isinstance(value, str) and value.strip().lower() in {
-                    "na",
-                    "n/a",
-                    "none",
-                    "null",
-                }:
-                    continue
-                info += f"<b>{label}:</b> {value}<br>"
-
-        self.content_info_text.setText(info if info else "No data available")
-
-        # Poster/cover image
-        poster_url = ""
-        if provider_type == "STB":
-            poster_url = item_data.get("screenshot_uri", "")
-        else:  # XTREAM and others
-            poster_url = item_data.get("logo") or item_data.get("cover") or ""
-
-        if poster_url:
-            self.lock_ui_before_loading()
-            if self.image_loader and self.image_loader.isRunning():
-                self.image_loader.wait()
-            self.image_loader = ImageLoader(
-                [poster_url],
-                self.image_manager,
-                iconified=False,
-                verify_ssl=self.config_manager.ssl_verify,
-            )
-            self.image_loader.progress_updated.connect(self.update_poster)
-            self.image_loader.finished.connect(self.image_loader_finished)
-            self.image_loader.start()
-            self.cancel_button.setText("Cancel fetching poster...")
-
-    def refresh_content_list_size(self):
-        font_size = 12
-        icon_size = font_size + 4
-        self.content_list.setIconSize(QSize(icon_size, icon_size))
-        self.content_list.setStyleSheet(
-            f"""
-        QTreeWidget {{ border: none; font-size: {font_size}px; }}
-        QTreeWidget::item {{ padding: 6px 8px; }}
-        """
-        )
-
-        font = QFont()
-        font.setPointSize(font_size)
-        self.content_list.setFont(font)
-
-        # Set header font
-        header_font = QFont()
-        header_font.setPointSize(font_size)
-        header_font.setBold(True)
-        self.content_list.header().setFont(header_font)
-
-    def toggle_favorite(self):
-        selected_item = self.content_list.currentItem()
-        if selected_item:
-            item_type = self.get_item_type(selected_item)
-            item_name = self.get_item_name(selected_item, item_type)
-            is_favorite = self.check_if_favorite(item_name)
-            if is_favorite:
-                self.remove_from_favorites(item_name)
-            else:
-                self.add_to_favorites(item_name)
-            self.filter_content(self.top_bar.search_text())
-
-    def add_to_favorites(self, item_name):
-        if item_name not in self.config_manager.favorites:
-            self.config_manager.favorites.append(item_name)
-            self.save_config()
-
-    def remove_from_favorites(self, item_name):
-        if item_name in self.config_manager.favorites:
-            self.config_manager.favorites.remove(item_name)
-            self.save_config()
-
-    def check_if_favorite(self, item_name):
-        return item_name in self.config_manager.favorites
-
-    def rescan_logos(self):
-        # Loop on content_list items to get logos and delete them from image_manager
-        logo_urls = []
-        for i in range(self.content_list.topLevelItemCount()):
-            item = self.content_list.topLevelItem(i)
-            url_logo = item.data(0, Qt.UserRole)["data"].get("logo", "")
-            logo_urls.append(url_logo)
-            if url_logo:
-                self.image_manager.remove_icon_from_cache(url_logo)
-
-        self.lock_ui_before_loading()
-        if self.image_loader and self.image_loader.isRunning():
-            self.image_loader.wait()
-        self.image_loader = ImageLoader(
-            logo_urls,
-            self.image_manager,
-            iconified=True,
-            verify_ssl=self.config_manager.ssl_verify,
-        )
-        self.image_loader.progress_updated.connect(self.update_channel_logos)
-        self.image_loader.finished.connect(self.image_loader_finished)
-        self.image_loader.start()
-        self.cancel_button.setText("Cancel fetching channel logos...")
-
-    def toggle_content_type(self, content_type=None):
-        """Switch content type. Called by sidebar or legacy code."""
-        if content_type:
-            self.content_type = content_type
-        self.current_category = None
-        self.current_series = None
-        self.current_season = None
-        self.navigation_stack.clear()
-        self.forward_stack.clear()
-        self.load_content()
-        self.top_bar.clear_search()
-
-    def display_categories(self, categories, select_first=True):
-        # Unregister the content_list selection change event
-        try:
-            self.content_list.itemSelectionChanged.disconnect(self.item_selected)
-        except (TypeError, RuntimeError):
-            pass
-        self.content_list.clear()
-        # Re-register the content_list selection change event
-        self.content_list.itemSelectionChanged.connect(self.item_selected)
-
-        # Stop refreshing content list
-        self.refresh_on_air_timer.stop()
-
-        self.current_list_content = "category"
-
-        self.content_list.setSortingEnabled(False)
-        self.content_list.setColumnCount(1)
-        if self.content_type == "itv":
-            self.content_list.setHeaderLabels([f"Channel Categories ({len(categories)})"])
-        elif self.content_type == "vod":
-            self.content_list.setHeaderLabels([f"Movie Categories ({len(categories)})"])
-        elif self.content_type == "series":
-            self.content_list.setHeaderLabels([f"Serie Categories ({len(categories)})"])
-
-        for category in categories:
-            item = CategoryTreeWidgetItem(self.content_list)
-            item.setText(0, category.get("title", "Unknown Category"))
-            item.setData(0, Qt.UserRole, {"type": "category", "data": category})
-            # Highlight favorite items
-            if self.check_if_favorite(category.get("title", "")):
-                item.setBackground(0, QColor(201, 107, 67, 24))
-
-        self.content_list.sortItems(0, Qt.AscendingOrder)
-        self.content_list.setSortingEnabled(True)
-        self.top_bar.set_back_visible(False)
-
-        self.clear_content_info_panel()
-
-        # Select an item in the list (first or a previously selected)
-        if select_first:
-            if select_first == True:
-                if self.content_list.topLevelItemCount() > 0:
-                    self.content_list.setCurrentItem(self.content_list.topLevelItem(0))
-            else:
-                previous_selected_id = select_first
-                previous_selected = self.content_list.findItems(
-                    previous_selected_id, Qt.MatchExactly, 0
-                )
-                if previous_selected:
-                    self.content_list.setCurrentItem(previous_selected[0])
-                    self.content_list.scrollToItem(previous_selected[0], QTreeWidget.PositionAtTop)
-
-    def display_content(self, items, content="m3ucontent", select_first=True):
-        # Stop refreshing On Air content BEFORE any structural changes
-        try:
-            if self.refresh_on_air_timer.isActive():
-                self.refresh_on_air_timer.stop()
-        except Exception:
-            pass
-
-        # Unregister the selection change event during rebuild
-        try:
-            self.content_list.itemSelectionChanged.disconnect(self.item_selected)
-        except (TypeError, RuntimeError):
-            pass
-
-        # Disable widget updates during clear (but keep signals enabled to allow Qt internal cleanup)
-        self.content_list.setUpdatesEnabled(False)
-
-        try:
-            self.content_list.clear()
-        except Exception as e:
-            logger.error(f"Error clearing content_list: {e}", exc_info=True)
-
-        try:
-            self.content_list.setSortingEnabled(False)
-        except (RuntimeError, AttributeError):
-            pass
-
-        # Defer reconnecting itemSelectionChanged until population completes
-
-        self.current_list_content = content
-        need_logos = content in ["channel", "m3ucontent"] and self.config_manager.channel_logos
-        logo_urls = []
-        use_epg = self.can_show_epg(content) and self.config_manager.channel_epg
-
-        # Define headers for different content types
-        category_header = self.current_category.get("title", "") if self.current_category else ""
-        serie_header = self.current_series.get("name", "") if self.current_series else ""
-        season_header = self.current_season.get("name", "") if self.current_season else ""
-        try:
-            serie_headers_str = f"{category_header} > Series ({len(items)})"
-            serie_headers_shortened = self.shorten_header(serie_headers_str)
-        except Exception as e:
-            logger.error(f"display_content: Error creating serie headers: {e}", exc_info=True)
-            serie_headers_shortened = "Series"
-
-        header_info = {
-            "serie": {
-                "headers": [
-                    serie_headers_shortened,
-                    "Genre",
-                    "Added",
-                ],
-                "keys": ["name", "genres_str", "added"],
-            },
-            "movie": {
-                "headers": [
-                    self.shorten_header(f"{category_header} > Movies ({len(items)})"),
-                    "Genre",
-                    "Added",
-                ],
-                "keys": ["name", "genres_str", "added"],
-            },
-            "season": {
-                "headers": [
-                    "#",
-                    (
-                        f"{category_header} > {serie_header} > Seasons"[:50]
-                        if len(f"{category_header} > {serie_header} > Seasons") > 50
-                        else f"{category_header} > {serie_header} > Seasons"
-                    ),
-                    "Added",
-                ],
-                "keys": ["number", "o_name", "added"],
-            },
-            "episode": {
-                "headers": [
-                    "#",
-                    self.shorten_header(
-                        f"{category_header} > {serie_header} > {season_header} > Episodes"
-                    ),
-                ],
-                "keys": ["number", "ename"],
-            },
-            "channel": {
-                "headers": [
-                    "#",
-                    self.shorten_header(f"{category_header} > Channels ({len(items)})"),
-                ]
-                + (["", "On Air"] if use_epg else []),
-                "keys": ["number", "name"],
-            },
-            "m3ucontent": {
-                "headers": [f"Name ({len(items)})", "Group"] + (["", "On Air"] if use_epg else []),
-                "keys": ["name", "group"],
-            },
-        }
-
-        # Get headers
-        headers = header_info[content]["headers"]
-        self.content_list.setColumnCount(len(headers))
-        try:
-            self.content_list.setHeaderLabels(headers)
-        except Exception as e:
-            logger.error(f"display_content: setHeaderLabels failed: {e}", exc_info=True)
-            raise
-
-        # no favorites on seasons or episodes genre_sfolders
-        check_fav = content in ["channel", "movie", "serie", "m3ucontent"]
-
-        # Disable updates during population to prevent Qt conflicts
-        self.content_list.setUpdatesEnabled(False)
-
-        for item_idx, item_data in enumerate(items):
-            # Create tree widget item based on content type
-            if content == "channel":
-                list_item = ChannelTreeWidgetItem(self.content_list)
-            elif content in ["season", "episode"]:
-                # Use NumberedTreeWidgetItem for seasons and episodes (numeric sorting)
-                list_item = NumberedTreeWidgetItem(self.content_list)
-            else:
-                # Use plain QTreeWidgetItem for other content
-                list_item = QTreeWidgetItem(self.content_list)
-
-            for col_idx, key in enumerate(header_info[content]["keys"]):
-                raw_value = item_data.get(key)
-                if key == "added":
-                    # Show only date part if present
-                    text_value = str(raw_value).split()[0] if raw_value else ""
-                else:
-                    text_value = html.unescape(str(raw_value)) if raw_value is not None else ""
-                list_item.setText(col_idx, text_value)
-
-            list_item.setData(0, Qt.UserRole, {"type": content, "data": item_data})
-
-            # If content type is channel, collect the logo urls from the image_manager
-            if need_logos:
-                logo_urls.append(item_data.get("logo", ""))
-
-            # Highlight favorite items
-            item_name = item_data.get("name") or item_data.get("title")
-            if check_fav and self.check_if_favorite(item_name):
-                list_item.setBackground(0, QColor(201, 107, 67, 24))
-
-        self.content_list.sortItems(0, Qt.AscendingOrder)
-        self.content_list.setSortingEnabled(True)
-
-        # Re-enable updates now that population and sorting are complete
-        self.content_list.setUpdatesEnabled(True)
-
-        # Resize columns AFTER re-enabling updates to avoid Qt timer conflicts
-        for i in range(len(header_info[content]["headers"])):
-            if i != 2:  # Don't auto-resize the progress column
-                try:
-                    self.content_list.resizeColumnToContents(i)
-                except Exception as e:
-                    logger.error(f"Error resizing column {i}: {e}", exc_info=True)
-
-        self.top_bar.set_back_visible(content != "m3ucontent")
-
-        if use_epg:
-            self.content_list.setItemDelegate(ChannelItemDelegate())
-            # Set a fixed width for the progress column
-            self.content_list.setColumnWidth(
-                2, 100
-            )  # Force column 2 (progress) to be 100 pixels wide
-            # Prevent user from resizing the progress column too small
-            self.content_list.header().setMinimumSectionSize(100)
-            # Start refreshing content list (currently aired program)
-            self.refresh_on_air()
-            self.refresh_on_air_timer.start(30000)
-
-        # Re-register the selection change event after rebuild
-        try:
-            self.content_list.itemSelectionChanged.connect(self.item_selected)
-        except Exception:
-            pass
-
-        # Select an item in the list (first or a previously selected)
-        if select_first:
-            if select_first == True:
-                if self.content_list.topLevelItemCount() > 0:
-                    self.content_list.setCurrentItem(self.content_list.topLevelItem(0))
-            else:
-                previous_selected_id = select_first
-                previous_selected = self.content_list.findItems(
-                    previous_selected_id, Qt.MatchExactly, 0
-                )
-                if previous_selected:
-                    self.content_list.setCurrentItem(previous_selected[0])
-                    self.content_list.scrollToItem(previous_selected[0], QTreeWidget.PositionAtTop)
-
-        # Load channel logos if needed
-        if need_logos:
-            self.lock_ui_before_loading()
-            if self.image_loader and self.image_loader.isRunning():
-                self.image_loader.wait()
-            self.image_loader = ImageLoader(
-                logo_urls,
-                self.image_manager,
-                iconified=True,
-                verify_ssl=self.config_manager.ssl_verify,
-            )
-            self.image_loader.progress_updated.connect(self.update_channel_logos)
-            self.image_loader.finished.connect(self.image_loader_finished)
-            self.image_loader.start()
-            self.cancel_button.setText("Cancel fetching channel logos...")
-
-    def update_channel_logos(self, current, total, data):
-        self.update_progress(current, total)
-        if data:
-            # Prefer using cache_path to construct GUI objects in the main thread
-            logo_column = ChannelList.get_logo_column(self.current_list_content)
-            rank = data.get("rank", 0)
-            item = (
-                self.content_list.topLevelItem(rank)
-                if rank < self.content_list.topLevelItemCount()
-                else None
-            )
-            if not item:
-                return
-            cache_path = data.get("cache_path")
-            if cache_path:
-                pix = QPixmap(cache_path)
-                if not pix.isNull():
-                    item.setIcon(logo_column, QIcon(pix))
-            else:
-                # Backward compatibility: if an icon was provided (older worker behavior)
-                qicon = data.get("icon", None)
-                if qicon:
-                    item.setIcon(logo_column, qicon)
-
-    def update_poster(self, current, total, data):
-        self.update_progress(current, total)
-        if data:
-            cache_path = data.get("cache_path")
-            pixmap = None
-            if cache_path:
-                pixmap = QPixmap(cache_path)
-            if pixmap and not pixmap.isNull():
-                scaled_pixmap = pixmap.scaled(200, 300, Qt.KeepAspectRatio, Qt.SmoothTransformation)
-                buffer = QBuffer()
-                buffer.open(QBuffer.ReadWrite)
-                scaled_pixmap.save(buffer, "PNG")
-                buffer.close()
-                base64_data = base64.b64encode(buffer.data()).decode("utf-8")
-                img_tag = f'<img src="data:image/png;base64,{base64_data}" alt="Poster Image" style="float:right; margin: 0 0 10px 10px;">'
-                self.content_info_text.setText(img_tag + self.content_info_text.text())
-
-    def filter_content(self, text=""):
-        # Cross-provider search mode
-        if getattr(self, "_all_providers_mode", False):
-            self._fusion_search(text)
-            return
-
-        show_favorites = self.sidebar.favorites_btn.isChecked()
-        search_text = text.lower() if isinstance(text, str) else ""
-
-        # When favorites is active at category level, switch to flat favorites view
-        if show_favorites and self.current_list_content == "category":
-            self._show_favorites_flat()
-            return
-
-        # retrieve items type first
-        item_type = None
-        if self.content_list.topLevelItemCount() > 0:
-            item = self.content_list.topLevelItem(0)
-            item_type = self.get_item_type(item)
-
-        for i in range(self.content_list.topLevelItemCount()):
-            item = self.content_list.topLevelItem(i)
-            item_name = self.get_item_name(item, item_type)
-            matches_search = search_text in item_name.lower()
-
-            # Optionally include metadata fields (description/plot, group, On Air EPG) in search
-            if (
-                not matches_search
-                and hasattr(self, "app_menu")
-                and self.app_menu.search_descriptions_action.isChecked()
-                and item_type in ["channel", "movie", "serie", "m3ucontent"]
-            ):
-                try:
-                    data = item.data(0, Qt.UserRole) or {}
-                    content = data.get("data", {}) if isinstance(data, dict) else {}
-                    description = content.get("description") or content.get("plot") or ""
-                    group = content.get("group", "")
-                    # Check textual metadata
-                    if (isinstance(description, str) and search_text in description.lower()) or (
-                        isinstance(group, str) and search_text in group.lower()
-                    ):
-                        matches_search = True
-                    # Check EPG "On Air" column text for channels/m3u when EPG is enabled
-                    if (
-                        not matches_search
-                        and item_type in ["channel", "m3ucontent"]
-                        and self.config_manager.channel_epg
-                        and self.can_show_epg(item_type)
-                    ):
-                        try:
-                            epg_text = item.data(3, Qt.UserRole) or ""
-                            if isinstance(epg_text, str) and search_text in epg_text.lower():
-                                matches_search = True
-                        except Exception:
-                            pass
-                except Exception:
-                    # Be conservative; ignore metadata if unexpected structure
-                    pass
-
-            # For categories, check if any content inside matches and show dropdown
-            if item_type == "category" and search_text:
-                matching_items = self._get_matching_items_in_category(item, search_text)
-                if matching_items:
-                    matches_search = True
-                    # Populate category with matching items as children
-                    self._populate_category_dropdown(item, matching_items)
-                    item.setExpanded(True)  # Auto-expand to show matches
-                else:
-                    # Category name matches but no content inside
-                    # Clear any existing children
-                    item.takeChildren()
-                    if not matches_search:
-                        item.setExpanded(False)
-            else:
-                # Not searching or not a category - clear children
-                if item_type == "category":
-                    item.takeChildren()
-                    item.setExpanded(False)
-
-            if item_type in ["category", "channel", "movie", "serie", "m3ucontent"]:
-                # For category, channel, movie, serie and generic content, filter by search text and favorite
-                is_favorite = self.check_if_favorite(item_name)
-                if show_favorites and not is_favorite:
-                    item.setHidden(True)
-                else:
-                    item.setHidden(not matches_search)
-            else:
-                # For season, episode, only filter by search text
-                item.setHidden(not matches_search)
-
-    def _fusion_search(self, text):
-        """Search across all providers' cached content."""
-        self.content_list.clear()
-
-        if not text or len(text) < 3:
-            self.content_list.setColumnCount(1)
-            self.content_list.setHeaderLabels(["Type to search across all providers..."])
-            return
-
-        search_text = text.lower()
-        self.content_list.setHeaderLabels([f'Search results for "{text}"'])
-        self.content_list.setSortingEnabled(False)
-        self.content_list.setColumnCount(1)
-
-        total_results = 0
-        type_to_item_type = {"itv": "channel", "vod": "movie", "series": "serie"}
-        source_cache = self._all_provider_cache_snapshot or []
-
-        for provider_name, cache in source_cache:
-            matches = []
-            for content_type in ["itv", "vod", "series"]:
-                content_data = cache.get(content_type, {})
-                if not isinstance(content_data, dict):
-                    continue
-                items = content_data.get("contents", content_data)
-                if isinstance(items, dict):
-                    # Could be category-based
-                    for cat_name, cat_items in items.items():
-                        if isinstance(cat_items, list):
-                            for item in cat_items:
-                                name = item.get("name", item.get("title", ""))
-                                if isinstance(name, str) and search_text in name.lower():
-                                    matches.append((content_type, item))
-                elif isinstance(items, list):
-                    for item in items:
-                        name = item.get("name", item.get("title", ""))
-                        if isinstance(name, str) and search_text in name.lower():
-                            matches.append((content_type, item))
-
-            if matches:
-                # Create provider group header styled as orange bar
-                provider_header = CategoryTreeWidgetItem(self.content_list)
-                provider_header.setText(0, f"  {provider_name} ({len(matches)} results)")
-                provider_header.setExpanded(True)
-                font = provider_header.font(0)
-                font.setBold(True)
-                provider_header.setFont(0, font)
-                provider_header.setBackground(0, QColor(201, 107, 67, 180))
-                provider_header.setForeground(0, QColor(255, 255, 255))
-
-                for content_type, item_data in matches[:50]:  # Limit per provider
-                    child = QTreeWidgetItem(provider_header)
-                    name = item_data.get("name", item_data.get("title", "Unknown"))
-                    type_prefix = {
-                        "itv": "[CH]",
-                        "vod": "[MOV]",
-                        "series": "[SER]",
-                    }.get(content_type, "")
-                    child.setText(0, f"{type_prefix} {name}")
-                    child.setData(
-                        0,
-                        Qt.UserRole,
-                        {
-                            "data": item_data,
-                            "type": type_to_item_type.get(content_type, "channel"),
-                            "provider": provider_name,
-                            "source_content_type": content_type,
-                            "cross_provider_result": True,
-                        },
-                    )
-                    total_results += 1
-
-        if total_results == 0:
-            self.content_list.setHeaderLabels([f'No results for "{text}"'])
-
-    def _get_matching_items_in_category(self, category_item, search_text):
-        """Get items in category that match the search text.
-
-        Returns list of matching items that can be displayed as dropdown children.
-        """
-        try:
-            data = category_item.data(0, Qt.UserRole)
-            if not data or "data" not in data:
-                return []
-
-            category_data = data["data"]
-            category_id = category_data.get("id", "*")
-
-            # Get provider content
-            content_data = self.provider_manager.current_provider_content.get(self.content_type, {})
-
-            # Check if we have categorized structure
-            if not isinstance(content_data, dict):
-                return []
-
-            # Get all items in this category
-            if category_id == "*":
-                # "All" category - check all contents
-                items = content_data.get("contents", [])
-            else:
-                # Specific category - get items by sorted_channels index
-                sorted_channels = content_data.get("sorted_channels", {})
-                indices = sorted_channels.get(category_id, [])
-                contents = content_data.get("contents", [])
-                items = [contents[i] for i in indices if i < len(contents)]
-
-            # Find matching items
-            search_lower = search_text.lower()
-            matching = []
-            include_desc = (
-                hasattr(self, "app_menu") and self.app_menu.search_descriptions_action.isChecked()
-            )
-            for item in items:
-                # Check name
-                name = item.get("name", "")
-                matches_name = search_lower in name.lower()
-
-                # Also check description for movies/series when enabled
-                description = item.get("description") or item.get("plot") or ""
-                matches_desc = include_desc and search_lower in description.lower()
-
-                # Optionally check current EPG program title when in live channels view
-                matches_epg = False
-                if (
-                    include_desc
-                    and self.content_type == "itv"
-                    and self.config_manager.channel_epg
-                    and self.can_show_epg("channel")
-                ):
-                    try:
-                        epg_list = self.epg_manager.get_programs_for_channel(item, None, 1) or []
-                        if epg_list:
-                            epg_item = epg_list[0]
-                            if "title" in epg_item:  # XMLTV style
-                                title_val = epg_item.get("title")
-                                epg_text = ""
-                                if isinstance(title_val, dict):
-                                    epg_text = title_val.get("__text") or ""
-                                elif isinstance(title_val, list) and title_val:
-                                    first = title_val[0]
-                                    if isinstance(first, dict):
-                                        epg_text = first.get("__text") or ""
-                                else:
-                                    epg_text = str(title_val or "")
-                            else:
-                                epg_text = str(epg_item.get("name") or "")
-                            matches_epg = search_lower in epg_text.lower()
-                    except Exception:
-                        matches_epg = False
-
-                if matches_name or matches_desc or matches_epg:
-                    matching.append(item)
-
-            return matching
-
-        except Exception as e:
-            # If anything goes wrong, return empty list
-            return []
-
-    def _populate_category_dropdown(self, category_item, items):
-        """Populate category tree item with matching content as children."""
-        # Clear existing children
-        category_item.takeChildren()
-
-        # Limit number of items shown in dropdown to avoid performance issues
-        max_items = 50
-        items_to_show = items[:max_items]
-        has_more = len(items) > max_items
-
-        # Add matching items as children
-        for item_data in items_to_show:
-            child_item = QTreeWidgetItem(category_item)
-            name = item_data.get("name", "Unknown")
-
-            # Add channel number if available
-            number = item_data.get("number", "")
-            if number:
-                child_item.setText(0, f"{number}. {name}")
-            else:
-                child_item.setText(0, name)
-
-            # Store item data
-            item_type = "channel"
-            if self.content_type == "vod":
-                item_type = "movie"
-            elif self.content_type == "series":
-                item_type = "serie"
-
-            child_item.setData(0, Qt.UserRole, {"type": item_type, "data": item_data})
-
-        # Add "... and X more" item if truncated
-        if has_more:
-            more_item = QTreeWidgetItem(category_item)
-            more_item.setText(
-                0, f"... and {len(items) - max_items} more (click category to see all)"
-            )
-            more_item.setDisabled(True)
-            # Gray out the text
-            font = more_item.font(0)
-            font.setItalic(True)
-            more_item.setFont(0, font)
-
-    def show_content_context_menu(self, position):
-        """Show context menu on right-click in content list."""
-        item = self.content_list.itemAt(position)
-        if not item:
-            return
-
-        item_data = item.data(0, Qt.UserRole)
-        if not item_data:
-            return
-
-        menu = QMenu(self)
-
-        # Favorite/Unfavorite action
-        item_type = self.get_item_type(item)
-        item_name = self.get_item_name(item, item_type)
-        is_fav = self.check_if_favorite(item_name)
-        fav_label = "\u2606 Remove from Favorites" if is_fav else "\u2605 Add to Favorites"
-        fav_action = menu.addAction(fav_label)
-        fav_action.triggered.connect(self.toggle_favorite)
-
-        # Copy URL action - only for playable content (not folders)
-        content_type = item_data.get("type", "")
-        if content_type not in ["category", "series"]:
-            url = self._get_stream_url_for_item(item_data)
-            if url:
-                menu.addSeparator()
-                copy_url_action = menu.addAction("Copy URL to Clipboard")
-                copy_url_action.triggered.connect(
-                    lambda checked=False, u=url: self._copy_url_to_clipboard(u)
-                )
-
-        menu.exec_(self.content_list.viewport().mapToGlobal(position))
-
-    def _get_stream_url_for_item(self, item_data):
-        """Get the stream URL for an item without playing it."""
-        # item_data is {"type": "...", "data": actual_item}
-        actual_data = item_data.get("data", {})
-        item_type = item_data.get("type", "")
-
-        if self.provider_manager.current_provider.get("type") == "STB":
-            # STB: need to create link via API
-            is_episode = item_type == "episode"
-            return self.create_link(actual_data, is_episode=is_episode)
-        else:
-            # M3U: cmd is the direct URL
-            return actual_data.get("cmd") or actual_data.get("url")
-
-    def _copy_url_to_clipboard(self, url):
-        """Copy URL to system clipboard."""
-        clipboard = QApplication.clipboard()
-        clipboard.setText(url)
-        logger.info(f"URL copied to clipboard: {url}")
-
-    def open_in_vlc(self):
-        # Invoke user's VLC player to open the current stream
-        if self.link:
-            logger.warning(f"Opening VLC for link: {self.link}")
-            try:
-                if platform.system() == "Windows":
-                    vlc_path = shutil.which("vlc")  # Try to find VLC in PATH
-                    if not vlc_path:
-                        program_files = os.environ.get("ProgramFiles", "C:\\Program Files")
-                        vlc_path = os.path.join(program_files, "VideoLAN", "VLC", "vlc.exe")
-                    # Use VLC's directory as cwd to avoid DLL conflicts with bundled libvlc
-                    vlc_dir = os.path.dirname(vlc_path)
-                    subprocess.Popen([vlc_path, self.link], cwd=vlc_dir)
-                elif platform.system() == "Darwin":  # macOS
-                    vlc_path = shutil.which("vlc")  # Try to find VLC in PATH
-                    if not vlc_path:
-                        common_paths = [
-                            "/Applications/VLC.app/Contents/MacOS/VLC",
-                            "~/Applications/VLC.app/Contents/MacOS/VLC",
-                        ]
-                        for path in common_paths:
-                            expanded_path = os.path.expanduser(path)
-                            if os.path.exists(expanded_path):
-                                vlc_path = expanded_path
-                                break
-                    if not vlc_path:
-                        raise FileNotFoundError("VLC not found")
-                    vlc_dir = os.path.dirname(vlc_path)
-                    subprocess.Popen([vlc_path, self.link], cwd=vlc_dir)
-                else:  # Assuming Linux or other Unix-like OS
-                    vlc_path = shutil.which("vlc")  # Try to find VLC in PATH
-                    if not vlc_path:
-                        raise FileNotFoundError("VLC not found")
-                    vlc_dir = os.path.dirname(vlc_path)
-                    subprocess.Popen([vlc_path, self.link], cwd=vlc_dir)
-                # when VLC opens, stop running video on self.player
-                self.player.stop_video()
-            except FileNotFoundError as fnf_error:
-                logger.warning("VLC not found: %s", fnf_error)
-            except Exception as e:
-                logger.warning(f"Error opening VLC: {e}")
-
-    def open_file(self):
-        file_path, _ = QFileDialog.getOpenFileName(self)
-        if file_path:
-            self._play_content(file_path)
-
-    def export_all_live_channels(self):
-        provider = self.provider_manager.current_provider
-        if provider.get("type") != "STB":
-            QMessageBox.warning(
-                self,
-                "Export Error",
-                "This feature is only available for STB providers.",
-            )
-            return
-
-        default_dir = str(Path.home() / "Downloads")
-        file_path, _ = QFileDialog.getSaveFileName(
-            self, "Export All Live Channels", default_dir, "M3U files (*.m3u)"
-        )
-        if file_path:
-            self.fetch_and_export_all_live_channels(file_path)
-
-    def fetch_and_export_all_live_channels(self, file_path):
-        selected_provider = self.provider_manager.current_provider
-        url = selected_provider.get("url", "")
-        url = URLObject(url)
-        base_url = f"{url.scheme}://{url.netloc}"
-        mac = selected_provider.get("mac", "")
-
-        try:
-            # Get all channels and categories (in provider cache)
-            provider_itv_content = self.provider_manager.current_provider_content.setdefault(
-                "itv", {}
-            )
-            categories_list = provider_itv_content.setdefault("categories", [])
-            categories = {
-                c.get("id", "None"): c.get("title", "Unknown Category") for c in categories_list
-            }
-            channels = provider_itv_content["contents"]
-
-            self.save_channel_list(base_url, channels, categories, mac, file_path)
-            QMessageBox.information(
-                self,
-                "Export Successful",
-                f"All live channels have been exported to {file_path}",
-            )
-        except Exception as e:
-            QMessageBox.critical(
-                self,
-                "Export Error",
-                f"An error occurred while exporting channels: {str(e)}",
-            )
-
-    def save_channel_list(self, base_url, channels_data, categories, mac, file_path) -> None:
-        try:
-            with open(file_path, "w", encoding="utf-8") as file:
-                file.write("#EXTM3U\n")
-                count = 0
-                for channel in channels_data:
-                    name = channel.get("name", "Unknown Channel")
-                    logo = channel.get("logo", "")
-                    category = channel.get("tv_genre_id", "None")
-                    xmltv_id = channel.get("xmltv_id", "")
-                    group = categories.get(category, "Unknown Group")
-                    cmd_url = channel.get("cmd", "").replace("ffmpeg ", "")
-                    if "localhost" in cmd_url:
-                        ch_id_match = re.search(r"/ch/(\d+)_", cmd_url)
-                        if ch_id_match:
-                            ch_id = ch_id_match.group(1)
-                            cmd_url = (
-                                f"{base_url}/play/live.php?mac={mac}&stream={ch_id}&extension=m3u8"
-                            )
-
-                    channel_str = f'#EXTINF:-1  tvg-id="{xmltv_id}" tvg-logo="{logo}" group-title="{group}" ,{name}\n{cmd_url}\n'
-                    count += 1
-                    file.write(channel_str)
-                logger.info(f"Channels = {count}")
-                logger.info(f"Channel list has been dumped to {file_path}")
-        except IOError as e:
-            logger.warning(f"Error saving channel list: {e}")
-
-    def export_shown_channels(self):
-        """Export the channels currently displayed in the list."""
-        items = []
-        for i in range(self.content_list.topLevelItemCount()):
-            tree_item = self.content_list.topLevelItem(i)
-            if tree_item:
-                user_data = tree_item.data(0, Qt.UserRole)
-                if user_data and "data" in user_data:
-                    items.append(user_data["data"])
-
-        if not items:
-            QMessageBox.warning(self, "Export", "No channels to export.")
-            return
-
-        default_dir = str(Path.home() / "Downloads")
-        file_path, _ = QFileDialog.getSaveFileName(
-            self, "Export Shown Channels", default_dir, "M3U files (*.m3u)"
-        )
-        if file_path:
-            provider = self.provider_manager.current_provider
-            config_type = provider.get("type", "")
-            if config_type == "STB":
-                base_url = provider.get("url", "")
-                mac = provider.get("mac", "")
-                save_stb_content(base_url, items, mac, file_path)
-            else:
-                save_m3u_content(items, file_path)
-
-    def export_content_cached(self):
-        """Export only the cached/browsed content that has already been loaded."""
-        default_dir = str(Path.home() / "Downloads")
-        file_path, _ = QFileDialog.getSaveFileName(
-            self, "Export Cached Content", default_dir, "M3U files (*.m3u)"
-        )
-        if file_path:
-            provider = self.provider_manager.current_provider
-            # Get the content data from the provider manager on content type
-            provider_content = self.provider_manager.current_provider_content.setdefault(
-                self.content_type, {}
-            )
-
-            base_url = provider.get("url", "")
-            config_type = provider.get("type", "")
-            mac = provider.get("mac", "")
-
-            if config_type == "STB":
-                # Extract all content items from categories
-                all_items = []
-                for items in provider_content.get("contents", {}).values():
-                    all_items.extend(items)
-                save_stb_content(base_url, all_items, mac, file_path)
-            elif config_type in ["M3UPLAYLIST", "M3USTREAM", "XTREAM"]:
-                content_items = provider_content if provider_content else []
-                save_m3u_content(content_items, file_path)
-            else:
-                logger.info(f"Unknown provider type: {config_type}")
-
-    def export_content_complete(self):
-        """Export all content by fetching all seasons/episodes for series (STB only)."""
-        provider = self.provider_manager.current_provider
-        config_type = provider.get("type", "")
-
-        # Check if this is appropriate content type
-        if config_type != "STB":
-            QMessageBox.information(
-                self,
-                "Export Complete",
-                "Complete export is only available for STB providers.\n\n"
-                "For other provider types, use 'Export Cached Content'.",
-            )
-            return
-
-        if self.content_type != "series":
-            if self.content_type == "itv":
-                QMessageBox.information(
-                    self,
-                    "Export Complete",
-                    "For live channels, please use 'Export All Live Channels' instead.\n\n"
-                    "Export Complete is designed for series with multiple seasons/episodes.",
-                )
-            else:
-                QMessageBox.information(
-                    self,
-                    "Export Complete",
-                    "Export Complete is only available for series content.\n\n"
-                    f"Current content type: {self.content_type}\n"
-                    "For movies or other content, use 'Export Cached Content'.",
-                )
-            return
-
-        default_dir = str(Path.home() / "Downloads")
-        file_path, _ = QFileDialog.getSaveFileName(
-            self,
-            "Export Complete Content (Fetch All)",
-            default_dir,
-            "M3U files (*.m3u)",
-        )
-        if file_path:
-            self.fetch_and_export_all_series(file_path)
-
-    def fetch_and_export_all_series(self, file_path):
-        """Fetch all series, seasons, and episodes, then export to M3U."""
-        selected_provider = self.provider_manager.current_provider
-        url = selected_provider.get("url", "")
-        url = URLObject(url)
-        base_url = f"{url.scheme}://{url.netloc}"
-        mac = selected_provider.get("mac", "")
-
-        # Get the current content (series in categories)
-        provider_content = self.provider_manager.current_provider_content.get(self.content_type, {})
-        categories = provider_content.get("categories", [])
-
-        if not categories:
-            QMessageBox.warning(
-                self,
-                "Export Error",
-                "No series categories found. Please load content first.",
-            )
-            return
-
-        # Show progress dialog
-        progress = QProgressDialog("Fetching all series data...", "Cancel", 0, 100, self)
-        progress.setWindowModality(Qt.WindowModal)
-        progress.setMinimumDuration(0)
-        progress.setValue(0)
-
-        all_episodes = []
-        total_series = 0
-
-        # Count total series across all categories
-        for cat_items in provider_content.get("contents", {}).values():
-            total_series += len(cat_items)
-
-        if total_series == 0:
-            progress.close()
-            QMessageBox.warning(
-                self,
-                "Export Error",
-                "No series found in loaded content.",
-            )
-            return
-
-        processed_series = 0
-
-        try:
-            # For each category
-            for category in categories:
-                category_id = category.get("id")
-                series_list = provider_content.get("contents", {}).get(category_id, [])
-
-                for series_item in series_list:
-                    if progress.wasCanceled():
-                        progress.close()
-                        return
-
-                    series_name = series_item.get("name", "Unknown")
-                    progress.setLabelText(f"Fetching: {series_name}")
-
-                    # Fetch seasons for this series
-                    seasons_data = self.fetch_seasons_sync(series_item)
-
-                    if seasons_data:
-                        seasons = seasons_data.get("data", [])
-                        for season in seasons:
-                            if progress.wasCanceled():
-                                progress.close()
-                                return
-
-                            # Fetch episodes for this season
-                            episodes_data = self.fetch_episodes_sync(series_item, season)
-
-                            if episodes_data:
-                                episodes = episodes_data.get("data", [])
-                                # Add series and season name to each episode for better identification
-                                for episode in episodes:
-                                    episode["series_name"] = series_name
-                                    episode["season_name"] = season.get("name", "")
-                                all_episodes.extend(episodes)
-
-                    processed_series += 1
-                    progress.setValue(int((processed_series / total_series) * 100))
-
-            progress.setValue(100)
-            progress.close()
-
-            # Now export all episodes
-            if all_episodes:
-                save_stb_content(base_url, all_episodes, mac, file_path)
-                QMessageBox.information(
-                    self,
-                    "Export Complete",
-                    f"Successfully exported {len(all_episodes)} episodes to {file_path}",
-                )
-            else:
-                QMessageBox.warning(
-                    self,
-                    "Export Warning",
-                    "No episodes found to export.",
-                )
-
-        except Exception as e:
-            progress.close()
-            logger.error(f"Error during complete export: {e}")
-            QMessageBox.critical(
-                self,
-                "Export Error",
-                f"An error occurred during export: {str(e)}",
-            )
-
-    def fetch_seasons_sync(self, series_item):
-        """Synchronously fetch seasons for a series."""
-        selected_provider = self.provider_manager.current_provider
-        headers = self.provider_manager.headers
-        url = selected_provider.get("url", "")
-        url = URLObject(url)
-        scheme = url.scheme
-        if self.config_manager.prefer_https and scheme == "http":
-            scheme = "https"
-        base_url = f"{scheme}://{url.netloc}/server/load.php"
-
-        params = {
-            "type": "series",
-            "action": "get_ordered_list",
-            "category_id": series_item.get("category_id"),
-            "movie_id": series_item.get("id"),
-            "season_id": 0,
-            "sortby": "name",
-            "JsHttpRequest": "1-xml",
-        }
-
-        try:
-            verify_ssl = selected_provider.get("ssl_verify", self.config_manager.ssl_verify)
-            response = requests.get(
-                base_url, headers=headers, params=params, timeout=10, verify=verify_ssl
-            )
-            if response.status_code == 200:
-                return response.json().get("js", {})
-        except Exception as e:
-            logger.warning(f"Error fetching seasons for {series_item.get('name')}: {e}")
-
-        return None
-
-    def fetch_episodes_sync(self, series_item, season_item):
-        """Synchronously fetch episodes for a season."""
-        selected_provider = self.provider_manager.current_provider
-        headers = self.provider_manager.headers
-        url = selected_provider.get("url", "")
-        url = URLObject(url)
-        scheme = url.scheme
-        if self.config_manager.prefer_https and scheme == "http":
-            scheme = "https"
-        base_url = f"{scheme}://{url.netloc}/server/load.php"
-
-        params = {
-            "type": "series",
-            "action": "get_ordered_list",
-            "category_id": series_item.get("category_id"),
-            "movie_id": series_item.get("id"),
-            "season_id": season_item.get("id"),
-            "sortby": "added",
-            "JsHttpRequest": "1-xml",
-        }
-
-        try:
-            verify_ssl = selected_provider.get("ssl_verify", self.config_manager.ssl_verify)
-            response = requests.get(
-                base_url, headers=headers, params=params, timeout=10, verify=verify_ssl
-            )
-            if response.status_code == 200:
-                return response.json().get("js", {})
-        except Exception as e:
-            logger.warning(
-                f"Error fetching episodes for {series_item.get('name')} - {season_item.get('name')}: {e}"
-            )
-
-        return None
-
-    # save_m3u_content and save_stb_content moved to services/export.py
-
-    def save_config(self):
-        self.config_manager.save_config()
-
-    def save_provider(self):
-        self.provider_manager.save_provider()
-
-    def load_content(self):
-        selected_provider = self.provider_manager.current_provider
-        config_type = selected_provider.get("type", "")
-        content = self.provider_manager.current_provider_content.setdefault(self.content_type, {})
-        if content:
-            # Check if content is a dict with categories (categorized format)
-            if isinstance(content, dict) and "categories" in content:
-                # Display categories for STB, XTREAM, and categorized M3U
-                self.display_categories(content.get("categories", []))
-            elif config_type in ("STB", "XTREAM"):
-                # Old cached format for STB/XTREAM - force update
-                self.update_content()
-            else:
-                # For flat M3U lists and other types, display content directly
-                self.display_content(content)
-        else:
-            self.update_content()
-
-    def update_content(self):
-        selected_provider = self.provider_manager.current_provider
-        config_type = selected_provider.get("type", "")
-        if config_type == "M3UPLAYLIST":
-            self.load_m3u_playlist(selected_provider["url"])
-        elif config_type == "XTREAM":
-            # Use Xtream Player API v2 with categories
-            self.load_xtream_content(
-                base_url=selected_provider.get("url", ""),
-                username=selected_provider.get("username", ""),
-                password=selected_provider.get("password", ""),
-                content_type=self.content_type,
-            )
-        elif config_type == "STB":
-            self.load_stb_categories(selected_provider["url"], self.provider_manager.headers)
-        elif config_type == "M3USTREAM":
-            self.load_stream(selected_provider["url"])
-
-    def load_m3u_playlist(self, url):
-        try:
-            if url.startswith(("http://", "https://")):
-                # Run network download in a worker thread
-                self.lock_ui_before_loading()
-                thread = QThread()
-                provider = self.provider_manager.current_provider
-                prefer_https = provider.get("prefer_https", self.config_manager.prefer_https)
-                verify_ssl = provider.get("ssl_verify", self.config_manager.ssl_verify)
-                worker = M3ULoaderWorker(url, verify_ssl=verify_ssl, prefer_https=prefer_https)
-                worker.moveToThread(thread)
-
-                def _cleanup():
-                    try:
-                        self._bg_jobs.remove((thread, worker))
-                    except ValueError:
-                        pass
-
-                # Start worker when thread starts
-                thread.started.connect(worker.run)
-                # Route results to UI thread
-                worker.finished.connect(self._on_m3u_loaded, Qt.QueuedConnection)
-                worker.error.connect(self._on_m3u_error, Qt.QueuedConnection)
-                # Ensure proper lifecycle
-                worker.finished.connect(thread.quit)
-                worker.error.connect(thread.quit)
-                thread.finished.connect(worker.deleteLater)
-                thread.finished.connect(thread.deleteLater)
-                thread.finished.connect(_cleanup)
-                thread.start()
-                self._bg_jobs.append((thread, worker))
-            else:
-                with open(url, "r", encoding="utf-8") as file:
-                    content = file.read()
-                # Parse with categorization enabled
-                parsed_content = parse_m3u(content, categorize=True)
-
-                # Check if we have categories (dict structure)
-                if isinstance(parsed_content, dict) and "categories" in parsed_content:
-                    # Store categorized content
-                    self.provider_manager.current_provider_content[self.content_type] = (
-                        parsed_content
-                    )
-                    self.save_provider()
-                    # Display categories
-                    self.display_categories(parsed_content.get("categories", []))
-                else:
-                    # Fallback to flat list
-                    self.display_content(parsed_content)
-                    self.provider_manager.current_provider_content[self.content_type] = (
-                        parsed_content
-                    )
-                    self.save_provider()
-        except (requests.RequestException, IOError) as e:
-            logger.warning(f"Error loading M3U Playlist: {e}")
-
-    def load_stream(self, url):
-        item = {"id": 1, "name": "Stream", "cmd": url}
-        self.display_content([item])
-        # Update the content in the config
-        self.provider_manager.current_provider_content[self.content_type] = [item]
-        self.save_provider()
+    # ------------------------------------------------------------------
+    # Item selection & activation
+    # ------------------------------------------------------------------
 
     def item_selected(self):
         if not self._info_panel_enabled:
@@ -3413,6 +1038,10 @@ class ChannelList(QMainWindow):
         if item_type in ["channel", "movie", "m3ucontent"]:
             self.play_item(item_data, item_type=item_type)
 
+    # ------------------------------------------------------------------
+    # Navigation
+    # ------------------------------------------------------------------
+
     def go_back(self):
         if self.navigation_stack:
             nav_type, previous_data, previous_selected_id = self.navigation_stack.pop()
@@ -3461,47 +1090,47 @@ class ChannelList(QMainWindow):
         finally:
             self._suppress_forward_clear = False
 
-    def options_dialog(self):
-        options = OptionsDialog(self)
-        options.exec_()
-        # Refresh provider combo in case providers were added/removed/renamed
-        self.populate_provider_combo()
+    def _is_playable_item(self, item: QTreeWidgetItem) -> bool:
+        try:
+            data = item.data(0, Qt.UserRole)
+            t = data.get("type") if isinstance(data, dict) else None
+            return t in {"m3ucontent", "channel", "movie", "episode"}
+        except Exception:
+            return False
 
-    # parse_m3u moved to services/m3u.py
+    def channel_surf_next(self):
+        """Move selection down by one; auto-play only if playable (not folders)."""
+        cl = self.content_list
+        count = cl.topLevelItemCount()
+        if count == 0:
+            return
+        current = cl.currentItem()
+        idx = cl.indexOfTopLevelItem(current) if current else -1
+        idx = (idx + 1) % count
+        candidate = cl.topLevelItem(idx)
+        if candidate is not None:
+            cl.setCurrentItem(candidate)
+            if self._is_playable_item(candidate):
+                self.item_activated(candidate)
 
-    def load_xtream_content(self, base_url: str, username: str, password: str, content_type: str):
-        """Load Xtream content (Live/VOD/Series) using Player API v2 with categories."""
-        self.lock_ui_before_loading()
-        thread = QThread()
-        provider = self.provider_manager.current_provider
-        prefer_https = provider.get("prefer_https", self.config_manager.prefer_https)
-        verify_ssl = provider.get("ssl_verify", self.config_manager.ssl_verify)
-        worker = XtreamLoaderWorker(
-            base_url,
-            username,
-            password,
-            content_type,
-            verify_ssl=verify_ssl,
-            prefer_https=prefer_https,
-        )
-        worker.moveToThread(thread)
+    def channel_surf_prev(self):
+        """Move selection up by one; auto-play only if playable (not folders)."""
+        cl = self.content_list
+        count = cl.topLevelItemCount()
+        if count == 0:
+            return
+        current = cl.currentItem()
+        idx = cl.indexOfTopLevelItem(current) if current else 0
+        idx = (idx - 1) % count
+        candidate = cl.topLevelItem(idx)
+        if candidate is not None:
+            cl.setCurrentItem(candidate)
+            if self._is_playable_item(candidate):
+                self.item_activated(candidate)
 
-        def _cleanup():
-            try:
-                self._bg_jobs.remove((thread, worker))
-            except ValueError:
-                pass
-
-        thread.started.connect(worker.run)
-        worker.finished.connect(self._on_xtream_content_loaded, Qt.QueuedConnection)
-        worker.error.connect(self._on_xtream_content_error, Qt.QueuedConnection)
-        worker.finished.connect(thread.quit)
-        worker.error.connect(thread.quit)
-        thread.finished.connect(worker.deleteLater)
-        thread.finished.connect(thread.deleteLater)
-        thread.finished.connect(_cleanup)
-        thread.start()
-        self._bg_jobs.append((thread, worker))
+    # ------------------------------------------------------------------
+    # Event handling
+    # ------------------------------------------------------------------
 
     def eventFilter(self, obj, event):
         try:
@@ -3536,945 +1165,25 @@ class ChannelList(QMainWindow):
             pass
         return super().eventFilter(obj, event)
 
-    def load_stb_categories(self, url: str, headers: Optional[dict] = None):
-        if headers is None:
-            headers = self.provider_manager.headers
-        # Run network calls in a worker thread
-        self.lock_ui_before_loading()
-        thread = QThread()
-        provider = self.provider_manager.current_provider
-        prefer_https = provider.get("prefer_https", self.config_manager.prefer_https)
-        verify_ssl = provider.get("ssl_verify", self.config_manager.ssl_verify)
-        worker = STBCategoriesWorker(
-            url,
-            headers,
-            self.content_type,
-            verify_ssl=verify_ssl,
-            prefer_https=prefer_https,
-        )
-        worker.moveToThread(thread)
-
-        def _cleanup():
-            try:
-                self._bg_jobs.remove((thread, worker))
-            except ValueError:
-                pass
-
-        thread.started.connect(worker.run)
-        worker.finished.connect(self._on_stb_categories_loaded, Qt.QueuedConnection)
-        worker.error.connect(self._on_stb_categories_error, Qt.QueuedConnection)
-        worker.finished.connect(thread.quit)
-        worker.error.connect(thread.quit)
-        thread.finished.connect(worker.deleteLater)
-        thread.finished.connect(thread.deleteLater)
-        thread.finished.connect(_cleanup)
-        thread.start()
-        self._bg_jobs.append((thread, worker))
-
-    @staticmethod
-    def get_categories_params(_type):
-        params = {
-            "type": _type,
-            "action": "get_genres" if _type == "itv" else "get_categories",
-            "JsHttpRequest": str(int(time.time() * 1000)) + "-xml",
-        }
-        return "&".join(f"{k}={v}" for k, v in params.items())
-
-    @staticmethod
-    def get_allchannels_params():
-        params = {
-            "type": "itv",
-            "action": "get_all_channels",
-            "JsHttpRequest": str(int(time.time() * 1000)) + "-xml",
-        }
-        return "&".join(f"{k}={v}" for k, v in params.items())
-
-    def load_content_in_category(self, category, select_first=True):
-        content_data = self.provider_manager.current_provider_content.setdefault(
-            self.content_type, {}
-        )
-        category_id = category.get("id", "*")
-        selected_provider = self.provider_manager.current_provider
-        config_type = selected_provider.get("type", "")
-
-        # For XTREAM and STB providers with sorted_channels structure
-        if "sorted_channels" in content_data:
-            contents = content_data.get("contents", [])
-            if category_id == "*":
-                items = contents if isinstance(contents, list) else []
-            else:
-                sorted_map = content_data.get("sorted_channels", {})
-                indices = sorted_map.get(category_id, []) if isinstance(sorted_map, dict) else []
-                # Guard missing/invalid state
-                items = [contents[i] for i in indices] if isinstance(contents, list) else []
-
-            # Display with appropriate content type
-            if self.content_type == "itv":
-                self.display_content(items, content="channel", select_first=select_first)
-            elif self.content_type == "series":
-                self.display_content(items, content="serie", select_first=select_first)
-            elif self.content_type == "vod":
-                self.display_content(items, content="movie", select_first=select_first)
-        else:
-            # For STB providers with per-category fetching
-            # Check if we have cached content for this category
-            if category_id in content_data.get("contents", {}):
-                items = content_data["contents"][category_id]
-                if self.content_type == "itv":
-                    self.display_content(items, content="channel", select_first=select_first)
-                elif self.content_type == "series":
-                    self.display_content(items, content="serie", select_first=select_first)
-                elif self.content_type == "vod":
-                    self.display_content(items, content="movie", select_first=select_first)
-            else:
-                # Fetch content for the category (STB only)
-                if config_type == "STB":
-                    self.fetch_content_in_category(category_id, select_first=select_first)
-
-    def fetch_content_in_category(self, category_id, select_first=True):
-        # Ask confirmation if the user wants to load all content
-        if category_id == "*":
-            reply = QMessageBox.question(
-                self,
-                "Load All Content",
-                "This will load all content in this category. Continue?",
-                QMessageBox.Yes | QMessageBox.No,
-                QMessageBox.No,
-            )
-            if reply == QMessageBox.No:
-                return
-
-        selected_provider = self.provider_manager.current_provider
-        headers = self.provider_manager.headers
-        url = selected_provider.get("url", "")
-        url = URLObject(url)
-        scheme = url.scheme
-        if (
-            selected_provider.get("prefer_https", self.config_manager.prefer_https)
-            and scheme == "http"
-        ):
-            scheme = "https"
-        url = f"{scheme}://{url.netloc}/server/load.php"
-
-        self.lock_ui_before_loading()
-        if self.content_loader and self.content_loader.isRunning():
-            self.content_loader.wait()
-        verify_ssl = selected_provider.get("ssl_verify", self.config_manager.ssl_verify)
-        self.content_loader = ContentLoader(
-            url,
-            headers,
-            self.content_type,
-            category_id=category_id,
-            verify_ssl=verify_ssl,
-        )
-        self.content_loader.content_loaded.connect(
-            lambda data: self.update_content_list(data, select_first)
-        )
-        self.content_loader.progress_updated.connect(self.update_progress)
-        self.content_loader.finished.connect(self.content_loader_finished)
-        self.content_loader.start()
-        self.cancel_button.setText("Cancel loading content in category")
-
-    def load_series_seasons(self, series_item, select_first=True):
-        selected_provider = self.provider_manager.current_provider
-        config_type = selected_provider.get("type", "")
-
-        self.current_series = series_item  # Store current series
-
-        if config_type == "XTREAM":
-            # Use Xtream API v2 to get series info
-            self.load_xtream_series_info(series_item, select_first)
-        elif config_type == "STB":
-            # Use STB API
-            headers = self.provider_manager.headers
-            url = selected_provider.get("url", "")
-            url = URLObject(url)
-            scheme = url.scheme
-            if (
-                selected_provider.get("prefer_https", self.config_manager.prefer_https)
-                and scheme == "http"
-            ):
-                scheme = "https"
-            url = f"{scheme}://{url.netloc}/server/load.php"
-
-            self.lock_ui_before_loading()
-            if self.content_loader and self.content_loader.isRunning():
-                self.content_loader.wait()
-            verify_ssl = selected_provider.get("ssl_verify", self.config_manager.ssl_verify)
-            self.content_loader = ContentLoader(
-                url=url,
-                headers=headers,
-                content_type="series",
-                category_id=series_item.get("category_id") or series_item.get("tv_genre_id", ""),
-                movie_id=series_item["id"],  # series ID
-                season_id=0,
-                action="get_ordered_list",
-                sortby="name",
-                verify_ssl=verify_ssl,
-            )
-
-            self.content_loader.content_loaded.connect(
-                lambda data: self.update_seasons_list(data, select_first)
-            )
-            self.content_loader.progress_updated.connect(self.update_progress)
-            self.content_loader.finished.connect(self.content_loader_finished)
-            self.content_loader.start()
-            self.cancel_button.setText("Cancel loading seasons")
-
-    def load_xtream_series_info(self, series_item, select_first=True):
-        """Load Xtream series info (seasons and episodes) using Player API v2."""
-        selected_provider = self.provider_manager.current_provider
-        content_data = self.provider_manager.current_provider_content.get(self.content_type, {})
-        resolved_base = content_data.get("resolved_base", "")
-
-        # Clean up any running background jobs before starting new one
-        for old_thread, old_worker in self._bg_jobs[
-            :
-        ]:  # Copy list to avoid modification during iteration
-            try:
-                if old_thread.isRunning():
-                    logger.info("Stopping previous worker thread")
-                    old_thread.quit()
-                    old_thread.wait(1000)  # Wait up to 1 second
-                    old_worker.deleteLater()
-                    old_thread.deleteLater()
-            except RuntimeError:
-                # Thread already deleted by Qt
-                pass
-            try:
-                self._bg_jobs.remove((old_thread, old_worker))
-            except ValueError:
-                # Already removed
-                pass
-
-        self.lock_ui_before_loading()
-        thread = QThread(self)
-        prefer_https = selected_provider.get("prefer_https", self.config_manager.prefer_https)
-        verify_ssl = selected_provider.get("ssl_verify", self.config_manager.ssl_verify)
-        worker = XtreamSeriesInfoWorker(
-            base_url=selected_provider.get("url", ""),
-            username=selected_provider.get("username", ""),
-            password=selected_provider.get("password", ""),
-            series_id=str(series_item["id"]),
-            resolved_base=resolved_base,
-            verify_ssl=verify_ssl,
-            prefer_https=prefer_https,
-        )
-        worker.moveToThread(thread)
-
-        # Remember selection behavior for UI handler
-        self._pending_series_select_first = select_first
-
-        # No nested UI handler; connect directly to a bound method
-
-        def on_error(msg):
-            # Route to UI handler
-            self._on_xtream_series_error(msg)
-
-        def _cleanup():
-            try:
-                self._bg_jobs.remove((thread, worker))
-            except ValueError:
-                pass
-
-        # Start worker when thread starts
-        thread.started.connect(worker.run)
-
-        # Route worker results to UI thread and ensure cleanup
-        worker.finished.connect(self._on_xtream_series_finished, Qt.QueuedConnection)
-        worker.error.connect(on_error, Qt.QueuedConnection)
-
-        # Ensure proper thread/worker lifecycle
-        worker.finished.connect(thread.quit)
-        worker.error.connect(thread.quit)
-        # Delete worker and thread objects in UI thread after finish
-        thread.finished.connect(worker.deleteLater)
-        thread.finished.connect(thread.deleteLater)
-        thread.finished.connect(_cleanup)
-        thread.start()
-        self._bg_jobs.append((thread, worker))
-
-    def _on_xtream_series_finished(self, payload):
-        """UI-thread handler for Xtream series info results."""
-        try:
-            if payload and isinstance(payload, dict):
-                seasons = payload.get("seasons", [])
-                if seasons:
-                    # Cache seasons for auto-play navigation
-                    self._current_seasons_list = seasons
-
-                    # Also cache in provider content for persistence
-                    if self.current_series:
-                        series_id = str(self.current_series.get("id", ""))
-                        content_data = self.provider_manager.current_provider_content.setdefault(
-                            "series", {}
-                        )
-                        seasons_cache = content_data.setdefault("seasons", {})
-                        seasons_cache[series_id] = seasons
-
-                        # Store resolved_base for episode URL construction
-                        content_data["resolved_base"] = payload.get("resolved_base", "")
-
-                    select_first = getattr(self, "_pending_series_select_first", True)
-                    self.update_seasons_list({"items": seasons}, select_first)
-            else:
-                logger.warning("Invalid payload from series worker")
-        except Exception as e:
-            logger.error(f"Series finished handler exception: {e}", exc_info=True)
-        finally:
-            # Ensure UI is unlocked even if cleanup signals race
-            self.unlock_ui_after_loading()
-
-    def _on_xtream_series_error(self, msg: str):
-        try:
-            logger.warning(f"Error loading Xtream series info: {msg}")
-        finally:
-            self.unlock_ui_after_loading()
-
-    def _on_m3u_loaded(self, payload):
-        try:
-            content = payload.get("content", "")
-            # Parse with categorization enabled
-            parsed_content = parse_m3u(content, categorize=True)
-
-            if isinstance(parsed_content, dict) and "categories" in parsed_content:
-                self.provider_manager.current_provider_content[self.content_type] = parsed_content
-                self.save_provider()
-                self.display_categories(parsed_content.get("categories", []))
-            else:
-                self.display_content(parsed_content)
-                self.provider_manager.current_provider_content[self.content_type] = parsed_content
-                self.save_provider()
-        finally:
-            self.unlock_ui_after_loading()
-
-    def _on_m3u_error(self, msg: str):
-        try:
-            logger.warning(f"Error loading M3U Playlist: {msg}")
-        finally:
-            self.unlock_ui_after_loading()
-
-    def _on_xtream_content_loaded(self, payload):
-        try:
-            categories = payload.get("categories", [])
-            contents = payload.get("contents", [])
-            sorted_channels = payload.get("sorted_channels", {})
-
-            if not categories and not contents:
-                logger.info("No content found.")
-                return
-
-            provider_content = self.provider_manager.current_provider_content.setdefault(
-                self.content_type, {}
-            )
-            provider_content["categories"] = categories
-            provider_content["contents"] = contents
-            provider_content["sorted_channels"] = sorted_channels
-
-            # Store stream settings for later use
-            provider_content["stream_base"] = payload.get("stream_base", "")
-            provider_content["stream_ext"] = payload.get("stream_ext", "")
-            provider_content["resolved_base"] = payload.get("resolved_base", "")
-
-            self.save_provider()
-            self.display_categories(categories)
-        finally:
-            self.unlock_ui_after_loading()
-
-    def _on_xtream_content_error(self, msg: str):
-        try:
-            logger.warning(f"Error loading Xtream content: {msg}")
-        finally:
-            self.unlock_ui_after_loading()
-
-    def _on_stb_categories_loaded(self, payload):
-        try:
-            categories = payload.get("categories", [])
-            if not categories:
-                logger.info("No categories found.")
-                return
-            provider_content = self.provider_manager.current_provider_content.setdefault(
-                self.content_type, {}
-            )
-            provider_content["categories"] = categories
-            provider_content["contents"] = {}
-
-            if self.content_type == "itv":
-                provider_content["contents"] = payload.get("all_channels", [])
-
-                sorted_channels: Dict[str, List[int]] = {}
-                for i in range(len(provider_content["contents"])):
-                    genre_id = provider_content["contents"][i]["tv_genre_id"]
-                    category_id = str(genre_id)
-                    if category_id not in sorted_channels:
-                        sorted_channels[category_id] = []
-                    sorted_channels[category_id].append(i)
-
-                for cat in sorted_channels:
-                    sorted_channels[cat].sort(
-                        key=lambda x: int(provider_content["contents"][x]["number"])
-                    )
-
-                if "None" in sorted_channels:
-                    categories.append({"id": "None", "title": "Unknown Category"})
-
-                provider_content["sorted_channels"] = sorted_channels
-
-            self.save_provider()
-            self.display_categories(categories)
-        finally:
-            self.unlock_ui_after_loading()
-
-    def _on_stb_categories_error(self, msg: str):
-        try:
-            logger.warning(f"Error loading STB categories: {msg}")
-        finally:
-            self.unlock_ui_after_loading()
-
-    def _on_link_created(self, payload):
-        try:
-            ctx = getattr(self, "_pending_link_ctx", None)
-            link = self.sanitize_url(payload.get("link", ""))
-            if link:
-                self.link = link
-                self._play_content(link)
-                if ctx:
-                    self.save_last_watched(ctx["item_data"], ctx["item_type"], link)
-            else:
-                logger.warning("Failed to create link.")
-        finally:
-            self.unlock_ui_after_loading()
-            self._pending_link_ctx = None
-
-    def _on_link_error(self, msg: str):
-        try:
-            logger.warning(f"Error creating link: {msg}")
-        finally:
-            self.unlock_ui_after_loading()
-
-    def load_season_episodes(self, season_item, select_first=True):
-        selected_provider = self.provider_manager.current_provider
-        config_type = selected_provider.get("type", "")
-
-        self.current_season = season_item  # Store current season
-
-        if config_type == "XTREAM":
-            # Episodes are already loaded with the season data
-            episodes = season_item.get("episodes", [])
-            if episodes:
-                # Format episodes for Xtream
-                content_data = self.provider_manager.current_provider_content.get(
-                    self.content_type, {}
-                )
-                stream_base = content_data.get("stream_base", "")
-                stream_ext = content_data.get("stream_ext", "ts")
-                username = selected_provider.get("username", "")
-                password = selected_provider.get("password", "")
-
-                formatted_episodes = []
-                for idx, ep in enumerate(episodes, start=1):
-                    episode_id = ep.get("id")
-                    # Use container_extension from episode data, fallback to stream_ext
-                    container_ext = ep.get("container_extension") or stream_ext
-                    cmd = f"{stream_base}/series/{username}/{password}/{episode_id}.{container_ext}"
-                    # Robust episode number extraction
-                    number_str = None
-                    for key in ("episode_num", "episode", "num"):
-                        val = ep.get(key)
-                        if val is not None:
-                            s = str(val)
-                            if s.isdigit():
-                                number_str = str(int(s))
-                                break
-                    if not number_str:
-                        title_text = ep.get("title") or ep.get("name") or ""
-                        m = re.search(r"\d+", str(title_text))
-                        if m:
-                            number_str = str(int(m.group(0)))
-                    if not number_str:
-                        # Try Roman numerals in title/name with conservative rules
-                        val = find_roman_token(title_text)
-                        if val is not None:
-                            number_str = str(val)
-                    if not number_str and episode_id is not None:
-                        s = str(episode_id)
-                        if s.isdigit():
-                            number_str = str(int(s))
-                    if not number_str:
-                        number_str = str(idx)
-
-                    formatted_ep = {
-                        "id": episode_id,
-                        "number": number_str,
-                        "ename": ep.get("title") or f"Episode {number_str}",
-                        "name": ep.get("title") or f"Episode {number_str}",
-                        "description": ep.get("info") or "",
-                        "cmd": cmd,
-                        "logo": (
-                            ep.get("info", {}).get("movie_image")
-                            if isinstance(ep.get("info"), dict)
-                            else ""
-                        ),
-                        "season": ep.get("season"),
-                        "episode_num": number_str,
-                        "container_extension": container_ext,
-                    }
-                    formatted_episodes.append(formatted_ep)
-
-                # Display episodes directly for Xtream
-                self.display_content(
-                    formatted_episodes, content="episode", select_first=select_first
-                )
-            else:
-                logger.info("No episodes found for this season.")
-        elif config_type == "STB":
-            # Use STB API
-            if not self.current_category or not self.current_series:
-                logger.warning("Current category/series not set when loading season episodes")
-                return
-
-            headers = self.provider_manager.headers
-            url = selected_provider.get("url", "")
-            url = URLObject(url)
-            url = f"{url.scheme}://{url.netloc}/server/load.php"
-
-            self.lock_ui_before_loading()
-            if self.content_loader and self.content_loader.isRunning():
-                self.content_loader.wait()
-            verify_ssl = selected_provider.get("ssl_verify", self.config_manager.ssl_verify)
-            self.content_loader = ContentLoader(
-                url=url,
-                headers=headers,
-                content_type="series",
-                category_id=self.current_category["id"],  # Category ID
-                movie_id=self.current_series["id"],  # Series ID
-                season_id=season_item["id"],  # Season ID
-                action="get_ordered_list",
-                sortby="added",
-                verify_ssl=verify_ssl,
-            )
-            self.content_loader.content_loaded.connect(
-                lambda data: self.update_episodes_list(data, select_first)
-            )
-            self.content_loader.progress_updated.connect(self.update_progress)
-            self.content_loader.finished.connect(self.content_loader_finished)
-            self.content_loader.start()
-            self.cancel_button.setText("Cancel loading episodes")
-
-    def play_item(self, item_data, is_episode=False, item_type=None):
-        # Track current playing item for auto-play
-        self._current_playing_item = item_data
-        self._current_playing_type = item_type
-
-        # Track episode context for auto-play
-        if item_type == "episode":
-            self._track_episode_context(item_data)
-        elif item_type == "movie":
-            self._track_movie_context(item_data)
-
-        # Generate content ID for position tracking
-        provider_name = self.provider_manager.current_provider.get("name", "")
-        self._current_content_id = self.config_manager._generate_content_id(
-            item_data, provider_name
-        )
-
-        # Add to watch history
-        self.config_manager.add_to_watch_history(
-            item_data, item_type or "channel", item_data.get("cmd", ""), provider_name
-        )
-
-        if self.provider_manager.current_provider["type"] == "STB":
-            # Create link in a worker thread, then play
-            selected_provider = self.provider_manager.current_provider
-            headers = self.provider_manager.headers
-            base_url = selected_provider.get("url", "")
-            cmd = item_data.get("cmd")
-            series_param = item_data.get("series") if is_episode else None
-
-            self.lock_ui_before_loading()
-            thread = QThread()
-            prefer_https = selected_provider.get("prefer_https", self.config_manager.prefer_https)
-            verify_ssl = selected_provider.get("ssl_verify", self.config_manager.ssl_verify)
-            worker = LinkCreatorWorker(
-                base_url=base_url,
-                headers=headers,
-                content_type=self.content_type,
-                cmd=cmd,
-                is_episode=is_episode,
-                series_param=series_param,
-                verify_ssl=verify_ssl,
-                prefer_https=prefer_https,
-            )
-            worker.moveToThread(thread)
-
-            # Store context for UI handler
-            self._pending_link_ctx = {
-                "item_data": item_data,
-                "item_type": item_type or "channel",
-                "is_episode": is_episode,
-            }
-
-            def _cleanup():
-                try:
-                    self._bg_jobs.remove((thread, worker))
-                except ValueError:
-                    pass
-
-            thread.started.connect(worker.run)
-            worker.finished.connect(self._on_link_created, Qt.QueuedConnection)
-            worker.error.connect(self._on_link_error, Qt.QueuedConnection)
-            worker.finished.connect(thread.quit)
-            worker.error.connect(thread.quit)
-            thread.finished.connect(worker.deleteLater)
-            thread.finished.connect(thread.deleteLater)
-            thread.finished.connect(_cleanup)
-            thread.start()
-            self._bg_jobs.append((thread, worker))
-        else:
-            cmd = item_data.get("cmd")
-            self.link = cmd
-            self._play_content_with_resume_check(cmd, item_data)
-            # Save last watched
-            self.save_last_watched(item_data, item_type or "m3ucontent", cmd)
-
-    def _track_episode_context(self, item_data: Dict[str, Any]):
-        """Track episode context for auto-play navigation."""
-        # Build episode list from current season
-        if self.current_season:
-            self._current_episode_list = self._build_episode_list_from_season(self.current_season)
-
-            # Find current episode index
-            current_num = str(item_data.get("number", ""))
-            for idx, ep in enumerate(self._current_episode_list):
-                if str(ep.get("number", "")) == current_num:
-                    self._current_episode_index = idx
-                    break
-
-        # Track seasons for next-season navigation
-        if self.current_series:
-            content_data = self.provider_manager.current_provider_content.get("series", {})
-            series_id = self.current_series.get("id")
-            if series_id:
-                cached_seasons = content_data.get("seasons", {}).get(str(series_id), [])
-                if cached_seasons:
-                    self._current_seasons_list = cached_seasons
-
-    def _track_movie_context(self, item_data: Dict[str, Any]):
-        """Track movie context for auto-play suggestions."""
-        # Get movies from current category
-        if self.current_category:
-            category_id = self.current_category.get("id")
-            content_data = self.provider_manager.current_provider_content.get("vod", {})
-            contents = content_data.get("contents", {})
-            if isinstance(contents, dict) and category_id in contents:
-                self._current_category_movies = contents[category_id]
-            elif isinstance(contents, list):
-                self._current_category_movies = contents
-
-    def _play_content_with_resume_check(self, url: str, item_data: Dict[str, Any]):
-        """Play content, checking for resume position first."""
-        # Check for saved position
-        saved_position = self.config_manager.get_playback_position(self._current_content_id)
-
-        if saved_position and saved_position > 5000:  # More than 5 seconds
-            # Check if near the end (don't resume if >90% watched)
-            positions = self.config_manager.playback_positions or {}
-            entry = positions.get(self._current_content_id, {})
-            duration = entry.get("duration", 0)
-
-            if duration > 0:
-                ratio = saved_position / duration
-                if ratio >= 0.90:
-                    # Near end, start from beginning
-                    self._play_content(url)
-                    return
-
-            # Show resume dialog
-            content_name = item_data.get("name") or item_data.get("ename") or "Content"
-
-            def on_resume():
-                self._play_content_with_position(url, saved_position)
-
-            def on_start_over():
-                self.config_manager.clear_playback_position(self._current_content_id)
-                self._play_content(url)
-
-            self._show_resume_dialog(content_name, saved_position, on_resume, on_start_over)
-        else:
-            self._play_content(url)
-
-    def _play_content_with_position(self, url: str, resume_position: int):
-        """Play content with resume position."""
-        if self.config_manager.play_in_vlc:
-            self._launch_vlc(url)
-        elif self.config_manager.play_in_mpv:
-            self._launch_mpv(url)
-        else:
-            provider_type = self.provider_manager.current_provider.get("type", "").upper()
-            if provider_type in ("STB", "XTREAM"):
-                is_live = self.content_type == "itv"
-            else:
-                is_live = None
-            self.player.play_video(
-                url,
-                is_live=is_live,
-                content_id=self._current_content_id,
-                resume_position=resume_position,
-            )
-
-    def save_last_watched(self, item_data, item_type, link):
-        """Save the last watched item to config"""
-        self.config_manager.last_watched = {
-            "item_data": item_data,
-            "item_type": item_type,
-            "link": link,
-            "timestamp": datetime.now().isoformat(),
-            "provider_name": self.provider_manager.current_provider.get("name", ""),
-        }
+    # ------------------------------------------------------------------
+    # Config / Options
+    # ------------------------------------------------------------------
+
+    def save_config(self):
         self.config_manager.save_config()
 
-    def resume_last_watched(self):
-        """Resume playing the last watched item"""
-        last_watched = self.config_manager.last_watched
-        if not last_watched:
-            QMessageBox.information(self, "No History", "No previously watched content found.")
-            return
+    def save_provider(self):
+        self.provider_manager.save_provider()
 
-        # Check if the provider matches
-        current_provider_name = self.provider_manager.current_provider.get("name", "")
-        if last_watched.get("provider_name") != current_provider_name:
-            reply = QMessageBox.question(
-                self,
-                "Different Provider",
-                f"Last watched content was from provider '{last_watched.get('provider_name')}'. Continue anyway?",
-                QMessageBox.Yes | QMessageBox.No,
-                QMessageBox.No,
-            )
-            if reply == QMessageBox.No:
-                return
-            # Switch provider automatically and resume after switching
-            target_name = last_watched.get("provider_name", "")
-            if target_name:
-                # If provider exists, switch to it
-                names = [p.get("name", "") for p in self.provider_manager.providers]
-                if target_name in names:
-                    self._pending_resume = last_watched
-                    self.config_manager.selected_provider_name = target_name
-                    self.config_manager.save_config()
-                    # Trigger provider switch; playback continues in set_provider_finished
-                    QTimer.singleShot(0, lambda: self.set_provider())
-                    return
+    def options_dialog(self):
+        options = OptionsDialog(self)
+        options.exec_()
+        # Refresh provider combo in case providers were added/removed/renamed
+        self.populate_provider_combo()
 
-        # Play the last watched item
-        item_data = last_watched.get("item_data")
-        item_type = last_watched.get("item_type")
-        is_episode = item_type == "episode"
-
-        # For STB providers, always recreate the link (tokens expire)
-        # For M3U/stream providers, can use stored link directly
-        current_provider_type = self.provider_manager.current_provider.get("type", "")
-        if current_provider_type == "STB":
-            # Recreate link with fresh token
-            self.play_item(item_data, is_episode=is_episode, item_type=item_type)
-        elif last_watched.get("link"):
-            # Use stored link for non-STB providers
-            self.link = last_watched["link"]
-            self._play_content(self.link)
-        else:
-            # Fallback: recreate the link
-            self.play_item(item_data, is_episode=is_episode, item_type=item_type)
-
-    def cancel_loading(self):
-        if self.content_loader and self.content_loader.isRunning():
-            self.content_loader.terminate()
-            self.content_loader.wait()
-            self.content_loader_finished()
-            QMessageBox.information(self, "Cancelled", "Content loading has been cancelled.")
-        elif self.image_loader and self.image_loader.isRunning():
-            self.image_loader.terminate()
-            self.image_loader.wait()
-            self.image_loader_finished()
-            self.image_manager.save_index()
-            QMessageBox.information(self, "Cancelled", "Image loading has been cancelled.")
-
-    def lock_ui_before_loading(self):
-        self.update_ui_on_loading(loading=True)
-
-    def unlock_ui_after_loading(self):
-        self.update_ui_on_loading(loading=False)
-
-    def update_ui_on_loading(self, loading):
-        self.top_bar.back_button.setEnabled(not loading)
-        self.progress_bar.setVisible(loading)
-        self.cancel_button.setVisible(loading)
-        if loading:
-            self.content_list.setSelectionMode(QListWidget.NoSelection)
-        else:
-            self.content_list.setSelectionMode(QListWidget.SingleSelection)
-
-    def content_loader_finished(self):
-        if self.content_loader:
-            self.content_loader.deleteLater()
-            self.content_loader = None
-        self.unlock_ui_after_loading()
-
-    def image_loader_finished(self):
-        if self.image_loader:
-            self.image_loader.deleteLater()
-            self.image_loader = None
-        self.unlock_ui_after_loading()
-
-    def update_content_list(self, data, select_first=True):
-        category_id = data.get("category_id")
-        items = data.get("items")
-
-        # Cache the items in config
-        selected_provider = self.provider_manager.current_provider_content
-        content_data = selected_provider.setdefault(self.content_type, {})
-        contents = content_data.setdefault("contents", {})
-        contents[category_id] = items
-        self.save_provider()
-
-        if self.content_type == "series":
-            self.display_content(items, content="serie", select_first=select_first)
-        elif self.content_type == "vod":
-            self.display_content(items, content="movie", select_first=select_first)
-        elif self.content_type == "itv":
-            self.display_content(items, content="channel", select_first=select_first)
-
-    def update_seasons_list(self, data, select_first=True):
-        if not self.current_series:
-            logger.warning("Current series not set when updating seasons list")
-            return
-        items = data.get("items")
-        if not items:
-            logger.warning("No items in data when updating seasons list")
-            return
-
-        for i, item in enumerate(items):
-            try:
-                # Store original name before modification
-                original_name = item.get("name", f"Season {i + 1}")
-                item["o_name"] = original_name
-
-                # Derive a numeric season index for sorting/display
-                number_str = None
-                # Prefer explicit id/season_number if numeric
-                sid = str(item.get("id", ""))
-                if sid.isdigit():
-                    number_str = str(int(sid))
-                else:
-                    s_num = str(item.get("season_number", ""))
-                    if s_num.isdigit():
-                        number_str = str(int(s_num))
-                # Fallback: first integer found in the name (e.g., "Season 1", "1. Sezon", "S02")
-                if not number_str:
-                    m = re.search(r"\d+", original_name)
-                    if m:
-                        number_str = str(int(m.group(0)))
-                # Fallback: Roman numeral detection (e.g., "Season IV", "Rocky II")
-                if not number_str:
-                    val = find_roman_token(original_name)
-                    if val is not None:
-                        number_str = str(val)
-                # Final fallback: position index
-                if not number_str:
-                    number_str = str(i + 1)
-                item["number"] = number_str
-
-                # Create combined name
-                series_name = self.current_series.get("name", "Unknown Series")
-                item["name"] = f"{series_name}.{original_name}"
-
-                # Add "added" field if not present (use air_date or empty)
-                if "added" not in item:
-                    item["added"] = item.get("air_date", "")
-
-            except Exception as e:
-                logger.error(f"Error processing season {i}: {e}", exc_info=True)
-                # Set defaults if processing fails
-                item["o_name"] = item.get("name", f"Season {i + 1}")
-                item["number"] = str(i + 1)
-                item["name"] = f"{self.current_series.get('name', 'Unknown')}.Season {i + 1}"
-                item["added"] = ""
-        self.display_content(items, content="season", select_first=select_first)
-
-    def update_episodes_list(self, data, select_first=True):
-        if not self.current_series:
-            logger.warning("Current series not set when updating episodes list")
-            return
-        items = data.get("items")
-        selected_season = None
-        for item in items:
-            if item.get("id") == data.get("season_id"):
-                selected_season = item
-                break
-
-        if selected_season:
-            episodes = selected_season.get("series", [])
-            episode_items = []
-            for episode_num in episodes:
-                # merge episode data with series data
-                episode_item = self.current_series.copy()
-                episode_item["number"] = f"{episode_num}"
-                episode_item["ename"] = f"Episode {episode_num}"
-                episode_item["cmd"] = selected_season.get("cmd")
-                episode_item["series"] = episode_num
-                episode_items.append(episode_item)
-            self.display_content(episode_items, content="episode", select_first=select_first)
-        else:
-            logger.info("Season not found in data.")
-
-    def update_progress(self, current, total):
-        if total:
-            progress_percentage = int((current / total) * 100)
-            self.progress_bar.setValue(progress_percentage)
-            if progress_percentage == 100:
-                self.progress_bar.setVisible(False)
-            else:
-                self.progress_bar.setVisible(True)
-
-    def update_busy_progress(self, msg):
-        self.cancel_button.setText(msg)
-
-    def create_link(self, item, is_episode=False):
-        try:
-            selected_provider = self.provider_manager.current_provider
-            headers = self.provider_manager.headers
-            url = selected_provider.get("url", "")
-            url = URLObject(url)
-            scheme = url.scheme
-            if self.config_manager.prefer_https and scheme == "http":
-                scheme = "https"
-            url = f"{scheme}://{url.netloc}"
-            cmd = item.get("cmd")
-            if is_episode:
-                # For episodes, we need to pass 'series' parameter
-                series_param = item.get("series")  # This should be the episode number
-                fetchurl = (
-                    f"{url}/server/load.php?type={'vod' if self.content_type == 'series' else self.content_type}&action=create_link"
-                    f"&cmd={url_quote(cmd)}&series={series_param}&JsHttpRequest=1-xml"
-                )
-            else:
-                fetchurl = (
-                    f"{url}/server/load.php?type={self.content_type}&action=create_link"
-                    f"&cmd={url_quote(cmd)}&JsHttpRequest=1-xml"
-                )
-            verify_ssl = selected_provider.get("ssl_verify", self.config_manager.ssl_verify)
-            response = requests.get(fetchurl, headers=headers, timeout=5, verify=verify_ssl)
-            if response.status_code != 200 or not response.content:
-                logger.warning(
-                    f"Error creating link: status code {response.status_code}, response content empty"
-                )
-                return None
-            result = response.json()
-            link = result["js"]["cmd"].split(" ")[-1]
-            link = self.sanitize_url(link)
-            self.link = link
-            return link
-        except Exception as e:
-            logger.warning(f"Error creating link: {e}")
-            return None
+    # ------------------------------------------------------------------
+    # Static utilities
+    # ------------------------------------------------------------------
 
     @staticmethod
     def sanitize_url(url):
@@ -4497,579 +1206,3 @@ class ChannelList(QMainWindow):
     @staticmethod
     def get_logo_column(item_type):
         return 0 if item_type == "m3ucontent" else 1
-
-    def _play_content(self, url):
-        """Play content in VLC, MPV, or built-in player based on checkbox state."""
-        if self.config_manager.play_in_vlc:
-            self._launch_vlc(url)
-        elif self.config_manager.play_in_mpv:
-            self._launch_mpv(url)
-        else:
-            # Use built-in player
-            # Determine is_live hint based on provider type
-            # STB/XTREAM APIs explicitly separate live (itv) from VOD content
-            # M3U playlists don't distinguish, so let VLC auto-detect
-            provider_type = self.provider_manager.current_provider.get("type", "").upper()
-            if provider_type in ("STB", "XTREAM"):
-                is_live = self.content_type == "itv"
-            else:
-                # M3UPLAYLIST, M3USTREAM, etc. - let VLC detect via seekability
-                is_live = None
-            self.player.play_video(url, is_live=is_live, content_id=self._current_content_id)
-
-    def _launch_vlc(self, url):
-        """Launch VLC with platform-specific handling."""
-        vlc_cmd = None
-
-        if platform.system() == "Darwin":
-            # macOS: use 'open -a VLC' which reuses existing instance
-            vlc_paths = [
-                "/Applications/VLC.app",
-                os.path.expanduser("~/Applications/VLC.app"),
-            ]
-            for path in vlc_paths:
-                if os.path.exists(path):
-                    try:
-                        subprocess.Popen(["open", "-a", path, url])
-                        return True
-                    except Exception as e:
-                        QMessageBox.warning(
-                            self, "VLC Launch Failed", f"Failed to launch VLC: {str(e)}"
-                        )
-                        return False
-
-        elif platform.system() == "Windows":
-            # Windows: find VLC executable
-            vlc_cmd = shutil.which("vlc")
-            if not vlc_cmd:
-                possible_paths = [
-                    os.path.join(
-                        os.environ.get("ProgramFiles", r"C:\Program Files"),
-                        "VideoLAN",
-                        "VLC",
-                        "vlc.exe",
-                    ),
-                    os.path.join(
-                        os.environ.get("ProgramFiles(x86)", r"C:\Program Files (x86)"),
-                        "VideoLAN",
-                        "VLC",
-                        "vlc.exe",
-                    ),
-                ]
-                for path in possible_paths:
-                    if os.path.exists(path):
-                        vlc_cmd = path
-                        break
-        else:
-            # Linux
-            vlc_cmd = shutil.which("vlc")
-
-        if not vlc_cmd:
-            QMessageBox.warning(
-                self,
-                "VLC Not Found",
-                "VLC Media Player is not installed.\n\n"
-                "Please install VLC:\n"
-                "• macOS: Download from https://www.videolan.org/vlc/\n"
-                "• Linux: Use your package manager (apt install vlc)\n"
-                "• Windows: Download from https://www.videolan.org/vlc/",
-            )
-            return False
-
-        try:
-            # --started-from-file sends URL to existing VLC instance (replaces current)
-            subprocess.Popen([vlc_cmd, "--started-from-file", url])
-            return True
-        except Exception as e:
-            QMessageBox.warning(self, "VLC Launch Failed", f"Failed to launch VLC: {str(e)}")
-            return False
-
-    def _launch_mpv(self, url):
-        """Launch MPV - try python-mpv first, fall back to subprocess."""
-        # Try python-mpv for single-instance behavior
-        try:
-            import mpv
-
-            if self._external_mpv_player is None:
-                self._external_mpv_player = mpv.MPV(
-                    input_default_bindings=True,
-                    input_vo_keyboard=True,
-                    osc=True,
-                )
-
-            self._external_mpv_player.play(url)
-            return True
-        except (ImportError, OSError, Exception):
-            # python-mpv not available or failed, fall back to subprocess
-            pass
-
-        # Subprocess fallback with platform-specific paths
-        mpv_cmd = None
-
-        if platform.system() == "Darwin":
-            # macOS: check app bundle and brew paths
-            macos_paths = [
-                "/Applications/mpv.app/Contents/MacOS/mpv",
-                os.path.expanduser("~/Applications/mpv.app/Contents/MacOS/mpv"),
-                "/opt/homebrew/bin/mpv",
-                "/usr/local/bin/mpv",
-            ]
-            for path in macos_paths:
-                if os.path.exists(path):
-                    mpv_cmd = path
-                    break
-            if not mpv_cmd:
-                mpv_cmd = shutil.which("mpv")
-
-        elif platform.system() == "Windows":
-            mpv_cmd = shutil.which("mpv")
-            if not mpv_cmd:
-                # Check Windows Registry App Paths (set by mpv-install.bat)
-                # Try both HKEY_LOCAL_MACHINE and HKEY_CURRENT_USER
-                try:
-                    import winreg
-
-                    for hkey in [winreg.HKEY_LOCAL_MACHINE, winreg.HKEY_CURRENT_USER]:  # type: ignore[attr-defined]
-                        try:
-                            with winreg.OpenKey(  # type: ignore[attr-defined]
-                                hkey,
-                                r"SOFTWARE\Microsoft\Windows\CurrentVersion\App Paths\mpv.exe",
-                            ) as key:
-                                reg_path = winreg.QueryValue(key, None)  # type: ignore[attr-defined]
-                                if reg_path and os.path.exists(reg_path):
-                                    mpv_cmd = reg_path
-                                    break
-                        except Exception:
-                            continue
-                except Exception:
-                    pass
-            if not mpv_cmd:
-                user_profile = os.environ.get("USERPROFILE", "")
-                possible_paths = [
-                    # Common installation paths
-                    os.path.join(
-                        os.environ.get("ProgramFiles", r"C:\Program Files"),
-                        "mpv",
-                        "mpv.exe",
-                    ),
-                    os.path.join(
-                        os.environ.get("ProgramFiles(x86)", r"C:\Program Files (x86)"),
-                        "mpv",
-                        "mpv.exe",
-                    ),
-                    os.path.join(os.environ.get("LOCALAPPDATA", ""), "Programs", "mpv", "mpv.exe"),
-                    # Scoop package manager
-                    os.path.join(user_profile, "scoop", "apps", "mpv", "current", "mpv.exe"),
-                    # Chocolatey package manager
-                    os.path.join(
-                        os.environ.get("ProgramData", r"C:\ProgramData"),
-                        "chocolatey",
-                        "bin",
-                        "mpv.exe",
-                    ),
-                    # Portable/extracted mpv (common locations)
-                    r"C:\mpv\mpv.exe",
-                    os.path.join(user_profile, "mpv", "mpv.exe"),
-                    os.path.join(user_profile, "Downloads", "mpv", "mpv.exe"),
-                    os.path.join(user_profile, "Desktop", "mpv", "mpv.exe"),
-                    # Winget default location
-                    os.path.join(
-                        os.environ.get("LOCALAPPDATA", ""),
-                        "Microsoft",
-                        "WinGet",
-                        "Packages",
-                        "mpv.net_Mpv.net",
-                        "mpv",
-                        "mpv.exe",
-                    ),
-                ]
-                for path in possible_paths:
-                    if path and os.path.exists(path):
-                        mpv_cmd = path
-                        break
-        else:
-            # Linux
-            mpv_cmd = shutil.which("mpv")
-
-        if not mpv_cmd:
-            QMessageBox.warning(
-                self,
-                "MPV Not Found",
-                "MPV Media Player is not installed.\n\n"
-                "Please install MPV:\n"
-                "• macOS: brew install mpv\n"
-                "• Linux: Use your package manager (apt install mpv)\n"
-                "• Windows: Download from https://mpv.io/installation/",
-            )
-            return False
-
-        try:
-            subprocess.Popen([mpv_cmd, url])
-            return True
-        except Exception as e:
-            QMessageBox.warning(self, "MPV Launch Failed", f"Failed to launch MPV: {str(e)}")
-            return False
-
-    # --- Auto-Play Methods ---
-
-    def on_media_ended(self):
-        """Handle media ended signal - trigger auto-play if enabled."""
-        # Don't auto-play if using external player
-        if self.config_manager.play_in_vlc or self.config_manager.play_in_mpv:
-            return
-
-        if not self._current_playing_item or not self._current_playing_type:
-            return
-
-        if self._current_playing_type == "episode":
-            if self.config_manager.auto_play_episodes:
-                self._show_next_episode_dialog()
-        elif self._current_playing_type == "movie":
-            if self.config_manager.auto_play_movies:
-                self._show_next_movie_dialog()
-
-    def on_position_changed(self, position_ms: int, duration_ms: int):
-        """Handle position changed signal - save playback position."""
-        if self._current_content_id and position_ms > 0 and duration_ms > 0:
-            self.config_manager.save_playback_position(
-                self._current_content_id, position_ms, duration_ms
-            )
-
-    def _show_next_episode_dialog(self):
-        """Show dialog for auto-playing next episode."""
-        next_episode = self._get_next_episode()
-        if not next_episode:
-            # Try next season
-            next_season_episode = self._get_first_episode_of_next_season()
-            if next_season_episode:
-                next_episode = next_season_episode
-            else:
-                # Series complete
-                series_name = (
-                    self.current_series.get("name", "Series") if self.current_series else "Series"
-                )
-                dialog = SeriesCompleteDialog(self, series_name)
-                dialog.show()
-                return
-
-        episode_name = next_episode.get("ename") or next_episode.get("name") or "Next Episode"
-
-        self._autoplay_dialog = EpisodeAutoPlayDialog(
-            self,
-            next_episode_name=episode_name,
-            countdown_seconds=self.config_manager.DEFAULT_OPTION_EPISODE_COUNTDOWN,
-        )
-        self._autoplay_dialog.countdownFinished.connect(
-            lambda: self._play_next_episode(next_episode)
-        )
-        self._autoplay_dialog.cancelled.connect(self._cancel_autoplay)
-        self._autoplay_dialog.show()
-
-    def _show_next_movie_dialog(self):
-        """Show dialog for suggesting next movie from same category."""
-        next_movie = self._get_next_movie_from_category()
-        if not next_movie:
-            category_name = self.current_category.get("title", "") if self.current_category else ""
-            dialog = NoCategoryMoviesDialog(self, category_name)
-            dialog.show()
-            return
-
-        movie_name = next_movie.get("name") or "Next Movie"
-        category_name = self.current_category.get("title", "") if self.current_category else ""
-
-        self._autoplay_dialog = MovieSuggestionDialog(
-            self,
-            movie_name=movie_name,
-            category_name=category_name,
-            countdown_seconds=self.config_manager.DEFAULT_OPTION_MOVIE_COUNTDOWN,
-        )
-        self._autoplay_dialog.countdownFinished.connect(lambda: self._play_next_movie(next_movie))
-        self._autoplay_dialog.cancelled.connect(self._cancel_autoplay)
-        self._autoplay_dialog.show()
-
-    def _cancel_autoplay(self):
-        """Cancel auto-play dialog."""
-        self._autoplay_dialog = None
-
-    def _get_next_episode(self) -> Optional[Dict[str, Any]]:
-        """Get the next episode in the current season."""
-        if self._current_episode_index < 0 or not self._current_episode_list:
-            return None
-
-        next_index = self._current_episode_index + 1
-        if next_index < len(self._current_episode_list):
-            return self._current_episode_list[next_index]
-        return None
-
-    def _get_first_episode_of_next_season(self) -> Optional[Dict[str, Any]]:
-        """Get the first episode of the next season."""
-        if not self.current_season or not self._current_seasons_list:
-            return None
-
-        # Find current season index
-        current_season_id = self.current_season.get("id")
-        current_season_idx = -1
-        for idx, season in enumerate(self._current_seasons_list):
-            if season.get("id") == current_season_id:
-                current_season_idx = idx
-                break
-
-        if current_season_idx < 0:
-            return None
-
-        next_season_idx = current_season_idx + 1
-        if next_season_idx >= len(self._current_seasons_list):
-            return None
-
-        # Load next season's episodes
-        next_season = self._current_seasons_list[next_season_idx]
-        episodes = next_season.get("episodes", [])
-        if not episodes:
-            return None
-
-        # Build episode item from first episode
-        first_ep = episodes[0]
-        provider_type = self.provider_manager.current_provider.get("type", "").upper()
-
-        if provider_type == "XTREAM":
-            episode_item = self._build_xtream_episode_item(first_ep, next_season)
-        else:
-            # STB episodes
-            episode_item = self._build_stb_episode_item(first_ep, next_season)
-
-        # Update current season for playback context
-        self.current_season = next_season
-        self._current_episode_list = self._build_episode_list_from_season(next_season)
-        self._current_episode_index = 0
-
-        return episode_item
-
-    def _build_xtream_episode_item(self, episode_data: dict, season_data: dict) -> Dict[str, Any]:
-        """Build an episode item dict for Xtream provider."""
-        provider = self.provider_manager.current_provider
-        resolved_base = self.provider_manager.current_provider_content.get("series", {}).get(
-            "resolved_base", ""
-        ) or provider.get("url", "")
-
-        username = provider.get("username", "")
-        password = provider.get("password", "")
-
-        ep_id = episode_data.get("id")
-        container_ext = episode_data.get("container_extension", "mkv")
-        cmd = f"{resolved_base}/series/{username}/{password}/{ep_id}.{container_ext}"
-
-        return {
-            "id": ep_id,
-            "name": episode_data.get("title") or f"Episode {episode_data.get('episode_num', '')}",
-            "ename": episode_data.get("title") or f"Episode {episode_data.get('episode_num', '')}",
-            "number": str(episode_data.get("episode_num", "")),
-            "cmd": cmd,
-            "series": episode_data.get("episode_num"),
-            "info": episode_data.get("info", {}),
-        }
-
-    def _build_stb_episode_item(self, episode_num: int, season_data: dict) -> Dict[str, Any]:
-        """Build an episode item dict for STB provider."""
-        if self.current_series:
-            episode_item = self.current_series.copy()
-        else:
-            episode_item = {}
-
-        episode_item["number"] = str(episode_num)
-        episode_item["ename"] = f"Episode {episode_num}"
-        episode_item["cmd"] = season_data.get("cmd")
-        episode_item["series"] = episode_num
-        return episode_item
-
-    def _build_episode_list_from_season(self, season_data: dict) -> List[Dict[str, Any]]:
-        """Build a list of episode items from season data."""
-        episodes = season_data.get("episodes", [])
-        provider_type = self.provider_manager.current_provider.get("type", "").upper()
-
-        episode_list = []
-        for ep in episodes:
-            if provider_type == "XTREAM":
-                episode_list.append(self._build_xtream_episode_item(ep, season_data))
-            else:
-                # STB uses simple episode numbers
-                if isinstance(ep, int):
-                    episode_list.append(self._build_stb_episode_item(ep, season_data))
-                elif isinstance(ep, dict):
-                    ep_num = ep.get("episode_num") or ep.get("series") or 1
-                    episode_list.append(self._build_stb_episode_item(int(ep_num), season_data))
-
-        return episode_list
-
-    def _play_next_episode(self, episode_data: Dict[str, Any]):
-        """Play the next episode."""
-        self._autoplay_dialog = None
-
-        # Update episode index
-        for idx, ep in enumerate(self._current_episode_list):
-            if ep.get("id") == episode_data.get("id") or (
-                ep.get("number") == episode_data.get("number")
-                and ep.get("cmd") == episode_data.get("cmd")
-            ):
-                self._current_episode_index = idx
-                break
-
-        # Play the episode
-        self.play_item(episode_data, is_episode=True, item_type="episode")
-
-    def _get_next_movie_from_category(self) -> Optional[Dict[str, Any]]:
-        """Get the next unwatched movie from the same category."""
-        if not self._current_category_movies or not self._current_playing_item:
-            return None
-
-        current_id = self._current_playing_item.get("id")
-        provider_name = self.provider_manager.current_provider.get("name", "")
-
-        # Find unwatched movies, excluding current
-        unwatched = []
-        for movie in self._current_category_movies:
-            if movie.get("id") == current_id:
-                continue
-            content_id = self.config_manager._generate_content_id(movie, provider_name)
-            status = self.config_manager.get_watched_status(content_id)
-            if status != "watched":
-                unwatched.append(movie)
-
-        if unwatched:
-            # Return the first unwatched movie
-            return unwatched[0]
-
-        # If all watched, just pick the next one after current
-        for idx, movie in enumerate(self._current_category_movies):
-            if movie.get("id") == current_id:
-                next_idx = (idx + 1) % len(self._current_category_movies)
-                if self._current_category_movies[next_idx].get("id") != current_id:
-                    return self._current_category_movies[next_idx]
-                break
-
-        return None
-
-    def _play_next_movie(self, movie_data: Dict[str, Any]):
-        """Play the next movie."""
-        self._autoplay_dialog = None
-        self.play_item(movie_data, item_type="movie")
-
-    def _format_position_time(self, ms: int) -> str:
-        """Format milliseconds as HH:MM:SS or MM:SS."""
-        seconds = ms // 1000
-        minutes, seconds = divmod(seconds, 60)
-        hours, minutes = divmod(minutes, 60)
-        if hours > 0:
-            return f"{hours:02d}:{minutes:02d}:{seconds:02d}"
-        return f"{minutes:02d}:{seconds:02d}"
-
-    def _show_resume_dialog(self, content_name: str, position_ms: int, on_resume, on_start_over):
-        """Show dialog for resuming playback."""
-        position_text = self._format_position_time(position_ms)
-
-        dialog = ResumeCountdownDialog(
-            self,
-            content_name=content_name,
-            resume_position_text=position_text,
-            countdown_seconds=self.config_manager.DEFAULT_OPTION_RESUME_COUNTDOWN,
-        )
-        dialog.countdownFinished.connect(on_resume)
-        dialog.resumeClicked.connect(on_resume)
-        dialog.startOverClicked.connect(on_start_over)
-        dialog.exec()
-
-    def show_watch_history(self):
-        """Show watch history dialog."""
-        history = self.config_manager.watch_history or []
-        if not history:
-            QMessageBox.information(self, "Watch History", "No watch history available.")
-            return
-
-        dialog = QDialog(self)
-        dialog.setWindowTitle("Watch History")
-        dialog.setMinimumSize(500, 400)
-
-        layout = QVBoxLayout(dialog)
-
-        history_list = QListWidget()
-        for entry in history:
-            name = entry.get("name", "Unknown")
-            item_type = entry.get("item_type", "")
-            provider = entry.get("provider_name", "")
-            timestamp = entry.get("timestamp", "")[:10]  # Just date
-
-            # Get progress info
-            content_id = entry.get("content_id", "")
-            status = self.config_manager.get_watched_status(content_id)
-            status_text = {"watched": "✓", "partial": "◐", "unwatched": ""}.get(status, "")
-
-            display_text = f"{status_text} {name} [{item_type}] - {provider} ({timestamp})"
-            list_item = QListWidgetItem(display_text)
-            list_item.setData(Qt.UserRole, entry)
-            history_list.addItem(list_item)
-
-        layout.addWidget(history_list)
-
-        # Buttons
-        button_layout = QHBoxLayout()
-
-        play_button = QPushButton("Play")
-
-        def on_play():
-            selected = history_list.selectedItems()
-            if selected:
-                entry = selected[0].data(Qt.UserRole)
-                dialog.accept()
-                self._play_from_history(entry)
-
-        play_button.clicked.connect(on_play)
-        button_layout.addWidget(play_button)
-
-        clear_button = QPushButton("Clear History")
-
-        def on_clear():
-            self.config_manager.watch_history = []
-            self.config_manager.save_config()
-            dialog.accept()
-
-        clear_button.clicked.connect(on_clear)
-        button_layout.addWidget(clear_button)
-
-        close_button = QPushButton("Close")
-        close_button.clicked.connect(dialog.reject)
-        button_layout.addWidget(close_button)
-
-        layout.addLayout(button_layout)
-
-        # Double-click to play
-        history_list.itemDoubleClicked.connect(lambda item: on_play())
-
-        dialog.exec()
-
-    def _play_from_history(self, entry: Dict[str, Any]):
-        """Play an item from watch history."""
-        item_data = entry.get("item_data")
-        item_type = entry.get("item_type")
-        provider_name = entry.get("provider_name", "")
-
-        # Check if we need to switch providers
-        current_provider_name = self.provider_manager.current_provider.get("name", "")
-        if provider_name and provider_name != current_provider_name:
-            # Check if provider exists
-            names = [p.get("name", "") for p in self.provider_manager.providers]
-            if provider_name in names:
-                self._pending_resume = entry
-                self.config_manager.selected_provider_name = provider_name
-                self.config_manager.save_config()
-                QTimer.singleShot(0, lambda: self.set_provider())
-                return
-            else:
-                QMessageBox.warning(
-                    self,
-                    "Provider Not Found",
-                    f"Provider '{provider_name}' is no longer available.",
-                )
-                return
-
-        # Play the item
-        is_episode = item_type == "episode"
-        self.play_item(item_data, is_episode=is_episode, item_type=item_type)
